renamed files and moved tests to accoriding fixtures
