--- gentest/fixtures/YGAlignBaseline.html
@@ -1,87 +1,0 @@
-<div id="align_baseline" style="width: 100px; height: 100px; flex-direction:row; align-items: baseline;">
-  <div style="width: 50px; height: 50px;"></div>
-  <div style="width: 50px; height: 20px;"></div>
-</div>
-
-<div id="align_baseline_child" style="width: 100px; height: 100px; flex-direction:row; align-items: baseline;">
-  <div style="width: 50px; height: 50px;"></div>
-  <div style="width: 50px; height: 20px;">
-    <div style="width: 50px; height: 10px;"></div>
-  </div>
-</div>
-
-<div id="align_baseline_child_top" style="width: 100px; height: 100px; flex-direction:row; align-items: baseline;">
-  <div style="width: 50px; height: 50px;top:10px;"></div>
-  <div style="width: 50px; height: 20px;">
-    <div style="width: 50px; height: 10px;"></div>
-  </div>
-</div>
-
-<div id="align_baseline_child_top2" style="width: 100px; height: 100px; flex-direction:row; align-items: baseline;">
-  <div style="width: 50px; height: 50px;"></div>
-  <div style="width: 50px; height: 20px;top:5px">
-    <div style="width: 50px; height: 10px;"></div>
-  </div>
-</div>
-
-<div id="align_baseline_double_nested_child" style="width: 100px; height: 100px; flex-direction:row; align-items: baseline;">
-  <div style="width: 50px; height: 50px;">
-    <div style="width: 50px; height: 20px;"></div>
-  </div>
-  <div style="width: 50px; height: 20px;">
-    <div style="width: 50px; height: 15px;"></div>
-  </div>
-</div>
-
-
-<div id="align_baseline_column" style="width: 100px; height: 100px; flex-direction:column; align-items: baseline;">
-  <div style="width: 50px; height: 50px;"></div>
-  <div style="width: 50px; height: 20px;"></div>
-</div>
-
-<div id="align_baseline_child_margin" style="width: 100px; height: 100px; flex-direction:row; align-items: baseline;">
-  <div style="width: 50px; height: 50px;margin: 5px;"></div>
-  <div style="width: 50px; height: 20px;">
-    <div style="width: 50px; height: 10px;margin: 1px;"></div>
-  </div>
-</div>
-
-<div id="align_baseline_multiline" style="width: 100px; height: 100px; flex-direction:row; align-items: baseline;flex-wrap:wrap;">
-  <div style="width: 50px; height: 50px;"></div>
-  <div style="width: 50px; height: 20px;">
-    <div style="width: 50px; height: 10px;"></div>
-  </div>
-  <div style="width: 50px; height: 20px;">
-    <div style="width: 50px; height: 10px;"></div>
-  </div>
-    <div style="width: 50px; height: 50px;"></div>
-</div>
-
-<div id="align_baseline_self" style="width: 100px; height: 100px; flex-direction:row;">
-  <div style="width: 50px; height: 50px;align-self: baseline;"></div>
-  <div style="width: 50px; height: 20px;align-self: baseline;">
-    <div style="width: 50px; height: 10px;"></div>
-  </div>
-</div>
-
-<div id="align_baseline_multiline_column" style="width: 100px; height: 100px; flex-direction:column; align-items: baseline;flex-wrap:wrap;">
-  <div style="width: 50px; height: 50px;"></div>
-  <div style="width: 50px; height: 50px;">
-    <div style="width: 50px; height: 10px;"></div>
-  </div>
-  <div style="width: 50px; height: 20px;">
-    <div style="width: 50px; height: 10px;"></div>
-  </div>
-    <div style="width: 50px; height: 20px;"></div>
-</div>
-
-<div id="align_baseline_multiline_row_and_column" style="width: 100px; height: 100px; flex-direction:row; align-items: baseline;flex-wrap:wrap;">
-  <div style="width: 50px; height: 50px;"></div>
-  <div style="width: 50px; height: 50px;flex-direction:column;">
-    <div style="width: 50px; height: 10px;"></div>
-  </div>
-  <div style="width: 50px; height: 20px;flex-direction:column;">
-    <div style="width: 50px; height: 10px;"></div>
-  </div>
-    <div style="width: 50px; height: 20px;"></div>
-</div>
