--- java/tests/com/facebook/yoga/YGAlignBaseline.java
@@ -1,910 +1,0 @@
-/**
- * Copyright (c) 2014-present, Facebook, Inc.
- * All rights reserved.
- *
- * This source code is licensed under the BSD-style license found in the
- * LICENSE file in the root directory of this source tree. An additional grant
- * of patent rights can be found in the PATENTS file in the same directory.
- */
-
- // @Generated by gentest/gentest.rb from gentest/fixtures/YGAlignBaseline.html
-
-package com.facebook.yoga;
-
-import org.junit.Test;
-
-import static org.junit.Assert.assertEquals;
-
-public class YGAlignBaseline {
-  @Test
-  public void test_align_baseline() {
-    final YogaNode root = new YogaNode();
-    root.setFlexDirection(YogaFlexDirection.ROW);
-    root.setAlignItems(YogaAlign.BASELINE);
-    root.setWidth(100f);
-    root.setHeight(100f);
-
-    final YogaNode root_child0 = new YogaNode();
-    root_child0.setWidth(50f);
-    root_child0.setHeight(50f);
-    root.addChildAt(root_child0, 0);
-
-    final YogaNode root_child1 = new YogaNode();
-    root_child1.setWidth(50f);
-    root_child1.setHeight(20f);
-    root.addChildAt(root_child1, 1);
-    root.setDirection(YogaDirection.LTR);
-    root.calculateLayout();
-
-    assertEquals(0f, root.getLayoutX(), 0.0f);
-    assertEquals(0f, root.getLayoutY(), 0.0f);
-    assertEquals(100f, root.getLayoutWidth(), 0.0f);
-    assertEquals(100f, root.getLayoutHeight(), 0.0f);
-
-    assertEquals(0f, root_child0.getLayoutX(), 0.0f);
-    assertEquals(0f, root_child0.getLayoutY(), 0.0f);
-    assertEquals(50f, root_child0.getLayoutWidth(), 0.0f);
-    assertEquals(50f, root_child0.getLayoutHeight(), 0.0f);
-
-    assertEquals(50f, root_child1.getLayoutX(), 0.0f);
-    assertEquals(30f, root_child1.getLayoutY(), 0.0f);
-    assertEquals(50f, root_child1.getLayoutWidth(), 0.0f);
-    assertEquals(20f, root_child1.getLayoutHeight(), 0.0f);
-
-    root.setDirection(YogaDirection.RTL);
-    root.calculateLayout();
-
-    assertEquals(0f, root.getLayoutX(), 0.0f);
-    assertEquals(0f, root.getLayoutY(), 0.0f);
-    assertEquals(100f, root.getLayoutWidth(), 0.0f);
-    assertEquals(100f, root.getLayoutHeight(), 0.0f);
-
-    assertEquals(50f, root_child0.getLayoutX(), 0.0f);
-    assertEquals(0f, root_child0.getLayoutY(), 0.0f);
-    assertEquals(50f, root_child0.getLayoutWidth(), 0.0f);
-    assertEquals(50f, root_child0.getLayoutHeight(), 0.0f);
-
-    assertEquals(0f, root_child1.getLayoutX(), 0.0f);
-    assertEquals(30f, root_child1.getLayoutY(), 0.0f);
-    assertEquals(50f, root_child1.getLayoutWidth(), 0.0f);
-    assertEquals(20f, root_child1.getLayoutHeight(), 0.0f);
-  }
-
-  @Test
-  public void test_align_baseline_child() {
-    final YogaNode root = new YogaNode();
-    root.setFlexDirection(YogaFlexDirection.ROW);
-    root.setAlignItems(YogaAlign.BASELINE);
-    root.setWidth(100f);
-    root.setHeight(100f);
-
-    final YogaNode root_child0 = new YogaNode();
-    root_child0.setWidth(50f);
-    root_child0.setHeight(50f);
-    root.addChildAt(root_child0, 0);
-
-    final YogaNode root_child1 = new YogaNode();
-    root_child1.setWidth(50f);
-    root_child1.setHeight(20f);
-    root.addChildAt(root_child1, 1);
-
-    final YogaNode root_child1_child0 = new YogaNode();
-    root_child1_child0.setWidth(50f);
-    root_child1_child0.setHeight(10f);
-    root_child1.addChildAt(root_child1_child0, 0);
-    root.setDirection(YogaDirection.LTR);
-    root.calculateLayout();
-
-    assertEquals(0f, root.getLayoutX(), 0.0f);
-    assertEquals(0f, root.getLayoutY(), 0.0f);
-    assertEquals(100f, root.getLayoutWidth(), 0.0f);
-    assertEquals(100f, root.getLayoutHeight(), 0.0f);
-
-    assertEquals(0f, root_child0.getLayoutX(), 0.0f);
-    assertEquals(0f, root_child0.getLayoutY(), 0.0f);
-    assertEquals(50f, root_child0.getLayoutWidth(), 0.0f);
-    assertEquals(50f, root_child0.getLayoutHeight(), 0.0f);
-
-    assertEquals(50f, root_child1.getLayoutX(), 0.0f);
-    assertEquals(40f, root_child1.getLayoutY(), 0.0f);
-    assertEquals(50f, root_child1.getLayoutWidth(), 0.0f);
-    assertEquals(20f, root_child1.getLayoutHeight(), 0.0f);
-
-    assertEquals(0f, root_child1_child0.getLayoutX(), 0.0f);
-    assertEquals(0f, root_child1_child0.getLayoutY(), 0.0f);
-    assertEquals(50f, root_child1_child0.getLayoutWidth(), 0.0f);
-    assertEquals(10f, root_child1_child0.getLayoutHeight(), 0.0f);
-
-    root.setDirection(YogaDirection.RTL);
-    root.calculateLayout();
-
-    assertEquals(0f, root.getLayoutX(), 0.0f);
-    assertEquals(0f, root.getLayoutY(), 0.0f);
-    assertEquals(100f, root.getLayoutWidth(), 0.0f);
-    assertEquals(100f, root.getLayoutHeight(), 0.0f);
-
-    assertEquals(50f, root_child0.getLayoutX(), 0.0f);
-    assertEquals(0f, root_child0.getLayoutY(), 0.0f);
-    assertEquals(50f, root_child0.getLayoutWidth(), 0.0f);
-    assertEquals(50f, root_child0.getLayoutHeight(), 0.0f);
-
-    assertEquals(0f, root_child1.getLayoutX(), 0.0f);
-    assertEquals(40f, root_child1.getLayoutY(), 0.0f);
-    assertEquals(50f, root_child1.getLayoutWidth(), 0.0f);
-    assertEquals(20f, root_child1.getLayoutHeight(), 0.0f);
-
-    assertEquals(0f, root_child1_child0.getLayoutX(), 0.0f);
-    assertEquals(0f, root_child1_child0.getLayoutY(), 0.0f);
-    assertEquals(50f, root_child1_child0.getLayoutWidth(), 0.0f);
-    assertEquals(10f, root_child1_child0.getLayoutHeight(), 0.0f);
-  }
-
-  @Test
-  public void test_align_baseline_child_top() {
-    final YogaNode root = new YogaNode();
-    root.setFlexDirection(YogaFlexDirection.ROW);
-    root.setAlignItems(YogaAlign.BASELINE);
-    root.setWidth(100f);
-    root.setHeight(100f);
-
-    final YogaNode root_child0 = new YogaNode();
-    root_child0.setPosition(YogaEdge.TOP, 10f);
-    root_child0.setWidth(50f);
-    root_child0.setHeight(50f);
-    root.addChildAt(root_child0, 0);
-
-    final YogaNode root_child1 = new YogaNode();
-    root_child1.setWidth(50f);
-    root_child1.setHeight(20f);
-    root.addChildAt(root_child1, 1);
-
-    final YogaNode root_child1_child0 = new YogaNode();
-    root_child1_child0.setWidth(50f);
-    root_child1_child0.setHeight(10f);
-    root_child1.addChildAt(root_child1_child0, 0);
-    root.setDirection(YogaDirection.LTR);
-    root.calculateLayout();
-
-    assertEquals(0f, root.getLayoutX(), 0.0f);
-    assertEquals(0f, root.getLayoutY(), 0.0f);
-    assertEquals(100f, root.getLayoutWidth(), 0.0f);
-    assertEquals(100f, root.getLayoutHeight(), 0.0f);
-
-    assertEquals(0f, root_child0.getLayoutX(), 0.0f);
-    assertEquals(10f, root_child0.getLayoutY(), 0.0f);
-    assertEquals(50f, root_child0.getLayoutWidth(), 0.0f);
-    assertEquals(50f, root_child0.getLayoutHeight(), 0.0f);
-
-    assertEquals(50f, root_child1.getLayoutX(), 0.0f);
-    assertEquals(40f, root_child1.getLayoutY(), 0.0f);
-    assertEquals(50f, root_child1.getLayoutWidth(), 0.0f);
-    assertEquals(20f, root_child1.getLayoutHeight(), 0.0f);
-
-    assertEquals(0f, root_child1_child0.getLayoutX(), 0.0f);
-    assertEquals(0f, root_child1_child0.getLayoutY(), 0.0f);
-    assertEquals(50f, root_child1_child0.getLayoutWidth(), 0.0f);
-    assertEquals(10f, root_child1_child0.getLayoutHeight(), 0.0f);
-
-    root.setDirection(YogaDirection.RTL);
-    root.calculateLayout();
-
-    assertEquals(0f, root.getLayoutX(), 0.0f);
-    assertEquals(0f, root.getLayoutY(), 0.0f);
-    assertEquals(100f, root.getLayoutWidth(), 0.0f);
-    assertEquals(100f, root.getLayoutHeight(), 0.0f);
-
-    assertEquals(50f, root_child0.getLayoutX(), 0.0f);
-    assertEquals(10f, root_child0.getLayoutY(), 0.0f);
-    assertEquals(50f, root_child0.getLayoutWidth(), 0.0f);
-    assertEquals(50f, root_child0.getLayoutHeight(), 0.0f);
-
-    assertEquals(0f, root_child1.getLayoutX(), 0.0f);
-    assertEquals(40f, root_child1.getLayoutY(), 0.0f);
-    assertEquals(50f, root_child1.getLayoutWidth(), 0.0f);
-    assertEquals(20f, root_child1.getLayoutHeight(), 0.0f);
-
-    assertEquals(0f, root_child1_child0.getLayoutX(), 0.0f);
-    assertEquals(0f, root_child1_child0.getLayoutY(), 0.0f);
-    assertEquals(50f, root_child1_child0.getLayoutWidth(), 0.0f);
-    assertEquals(10f, root_child1_child0.getLayoutHeight(), 0.0f);
-  }
-
-  @Test
-  public void test_align_baseline_child_top2() {
-    final YogaNode root = new YogaNode();
-    root.setFlexDirection(YogaFlexDirection.ROW);
-    root.setAlignItems(YogaAlign.BASELINE);
-    root.setWidth(100f);
-    root.setHeight(100f);
-
-    final YogaNode root_child0 = new YogaNode();
-    root_child0.setWidth(50f);
-    root_child0.setHeight(50f);
-    root.addChildAt(root_child0, 0);
-
-    final YogaNode root_child1 = new YogaNode();
-    root_child1.setPosition(YogaEdge.TOP, 5f);
-    root_child1.setWidth(50f);
-    root_child1.setHeight(20f);
-    root.addChildAt(root_child1, 1);
-
-    final YogaNode root_child1_child0 = new YogaNode();
-    root_child1_child0.setWidth(50f);
-    root_child1_child0.setHeight(10f);
-    root_child1.addChildAt(root_child1_child0, 0);
-    root.setDirection(YogaDirection.LTR);
-    root.calculateLayout();
-
-    assertEquals(0f, root.getLayoutX(), 0.0f);
-    assertEquals(0f, root.getLayoutY(), 0.0f);
-    assertEquals(100f, root.getLayoutWidth(), 0.0f);
-    assertEquals(100f, root.getLayoutHeight(), 0.0f);
-
-    assertEquals(0f, root_child0.getLayoutX(), 0.0f);
-    assertEquals(0f, root_child0.getLayoutY(), 0.0f);
-    assertEquals(50f, root_child0.getLayoutWidth(), 0.0f);
-    assertEquals(50f, root_child0.getLayoutHeight(), 0.0f);
-
-    assertEquals(50f, root_child1.getLayoutX(), 0.0f);
-    assertEquals(45f, root_child1.getLayoutY(), 0.0f);
-    assertEquals(50f, root_child1.getLayoutWidth(), 0.0f);
-    assertEquals(20f, root_child1.getLayoutHeight(), 0.0f);
-
-    assertEquals(0f, root_child1_child0.getLayoutX(), 0.0f);
-    assertEquals(0f, root_child1_child0.getLayoutY(), 0.0f);
-    assertEquals(50f, root_child1_child0.getLayoutWidth(), 0.0f);
-    assertEquals(10f, root_child1_child0.getLayoutHeight(), 0.0f);
-
-    root.setDirection(YogaDirection.RTL);
-    root.calculateLayout();
-
-    assertEquals(0f, root.getLayoutX(), 0.0f);
-    assertEquals(0f, root.getLayoutY(), 0.0f);
-    assertEquals(100f, root.getLayoutWidth(), 0.0f);
-    assertEquals(100f, root.getLayoutHeight(), 0.0f);
-
-    assertEquals(50f, root_child0.getLayoutX(), 0.0f);
-    assertEquals(0f, root_child0.getLayoutY(), 0.0f);
-    assertEquals(50f, root_child0.getLayoutWidth(), 0.0f);
-    assertEquals(50f, root_child0.getLayoutHeight(), 0.0f);
-
-    assertEquals(0f, root_child1.getLayoutX(), 0.0f);
-    assertEquals(45f, root_child1.getLayoutY(), 0.0f);
-    assertEquals(50f, root_child1.getLayoutWidth(), 0.0f);
-    assertEquals(20f, root_child1.getLayoutHeight(), 0.0f);
-
-    assertEquals(0f, root_child1_child0.getLayoutX(), 0.0f);
-    assertEquals(0f, root_child1_child0.getLayoutY(), 0.0f);
-    assertEquals(50f, root_child1_child0.getLayoutWidth(), 0.0f);
-    assertEquals(10f, root_child1_child0.getLayoutHeight(), 0.0f);
-  }
-
-  @Test
-  public void test_align_baseline_double_nested_child() {
-    final YogaNode root = new YogaNode();
-    root.setFlexDirection(YogaFlexDirection.ROW);
-    root.setAlignItems(YogaAlign.BASELINE);
-    root.setWidth(100f);
-    root.setHeight(100f);
-
-    final YogaNode root_child0 = new YogaNode();
-    root_child0.setWidth(50f);
-    root_child0.setHeight(50f);
-    root.addChildAt(root_child0, 0);
-
-    final YogaNode root_child0_child0 = new YogaNode();
-    root_child0_child0.setWidth(50f);
-    root_child0_child0.setHeight(20f);
-    root_child0.addChildAt(root_child0_child0, 0);
-
-    final YogaNode root_child1 = new YogaNode();
-    root_child1.setWidth(50f);
-    root_child1.setHeight(20f);
-    root.addChildAt(root_child1, 1);
-
-    final YogaNode root_child1_child0 = new YogaNode();
-    root_child1_child0.setWidth(50f);
-    root_child1_child0.setHeight(15f);
-    root_child1.addChildAt(root_child1_child0, 0);
-    root.setDirection(YogaDirection.LTR);
-    root.calculateLayout();
-
-    assertEquals(0f, root.getLayoutX(), 0.0f);
-    assertEquals(0f, root.getLayoutY(), 0.0f);
-    assertEquals(100f, root.getLayoutWidth(), 0.0f);
-    assertEquals(100f, root.getLayoutHeight(), 0.0f);
-
-    assertEquals(0f, root_child0.getLayoutX(), 0.0f);
-    assertEquals(0f, root_child0.getLayoutY(), 0.0f);
-    assertEquals(50f, root_child0.getLayoutWidth(), 0.0f);
-    assertEquals(50f, root_child0.getLayoutHeight(), 0.0f);
-
-    assertEquals(0f, root_child0_child0.getLayoutX(), 0.0f);
-    assertEquals(0f, root_child0_child0.getLayoutY(), 0.0f);
-    assertEquals(50f, root_child0_child0.getLayoutWidth(), 0.0f);
-    assertEquals(20f, root_child0_child0.getLayoutHeight(), 0.0f);
-
-    assertEquals(50f, root_child1.getLayoutX(), 0.0f);
-    assertEquals(5f, root_child1.getLayoutY(), 0.0f);
-    assertEquals(50f, root_child1.getLayoutWidth(), 0.0f);
-    assertEquals(20f, root_child1.getLayoutHeight(), 0.0f);
-
-    assertEquals(0f, root_child1_child0.getLayoutX(), 0.0f);
-    assertEquals(0f, root_child1_child0.getLayoutY(), 0.0f);
-    assertEquals(50f, root_child1_child0.getLayoutWidth(), 0.0f);
-    assertEquals(15f, root_child1_child0.getLayoutHeight(), 0.0f);
-
-    root.setDirection(YogaDirection.RTL);
-    root.calculateLayout();
-
-    assertEquals(0f, root.getLayoutX(), 0.0f);
-    assertEquals(0f, root.getLayoutY(), 0.0f);
-    assertEquals(100f, root.getLayoutWidth(), 0.0f);
-    assertEquals(100f, root.getLayoutHeight(), 0.0f);
-
-    assertEquals(50f, root_child0.getLayoutX(), 0.0f);
-    assertEquals(0f, root_child0.getLayoutY(), 0.0f);
-    assertEquals(50f, root_child0.getLayoutWidth(), 0.0f);
-    assertEquals(50f, root_child0.getLayoutHeight(), 0.0f);
-
-    assertEquals(0f, root_child0_child0.getLayoutX(), 0.0f);
-    assertEquals(0f, root_child0_child0.getLayoutY(), 0.0f);
-    assertEquals(50f, root_child0_child0.getLayoutWidth(), 0.0f);
-    assertEquals(20f, root_child0_child0.getLayoutHeight(), 0.0f);
-
-    assertEquals(0f, root_child1.getLayoutX(), 0.0f);
-    assertEquals(5f, root_child1.getLayoutY(), 0.0f);
-    assertEquals(50f, root_child1.getLayoutWidth(), 0.0f);
-    assertEquals(20f, root_child1.getLayoutHeight(), 0.0f);
-
-    assertEquals(0f, root_child1_child0.getLayoutX(), 0.0f);
-    assertEquals(0f, root_child1_child0.getLayoutY(), 0.0f);
-    assertEquals(50f, root_child1_child0.getLayoutWidth(), 0.0f);
-    assertEquals(15f, root_child1_child0.getLayoutHeight(), 0.0f);
-  }
-
-  @Test
-  public void test_align_baseline_column() {
-    final YogaNode root = new YogaNode();
-    root.setAlignItems(YogaAlign.BASELINE);
-    root.setWidth(100f);
-    root.setHeight(100f);
-
-    final YogaNode root_child0 = new YogaNode();
-    root_child0.setWidth(50f);
-    root_child0.setHeight(50f);
-    root.addChildAt(root_child0, 0);
-
-    final YogaNode root_child1 = new YogaNode();
-    root_child1.setWidth(50f);
-    root_child1.setHeight(20f);
-    root.addChildAt(root_child1, 1);
-    root.setDirection(YogaDirection.LTR);
-    root.calculateLayout();
-
-    assertEquals(0f, root.getLayoutX(), 0.0f);
-    assertEquals(0f, root.getLayoutY(), 0.0f);
-    assertEquals(100f, root.getLayoutWidth(), 0.0f);
-    assertEquals(100f, root.getLayoutHeight(), 0.0f);
-
-    assertEquals(0f, root_child0.getLayoutX(), 0.0f);
-    assertEquals(0f, root_child0.getLayoutY(), 0.0f);
-    assertEquals(50f, root_child0.getLayoutWidth(), 0.0f);
-    assertEquals(50f, root_child0.getLayoutHeight(), 0.0f);
-
-    assertEquals(0f, root_child1.getLayoutX(), 0.0f);
-    assertEquals(50f, root_child1.getLayoutY(), 0.0f);
-    assertEquals(50f, root_child1.getLayoutWidth(), 0.0f);
-    assertEquals(20f, root_child1.getLayoutHeight(), 0.0f);
-
-    root.setDirection(YogaDirection.RTL);
-    root.calculateLayout();
-
-    assertEquals(0f, root.getLayoutX(), 0.0f);
-    assertEquals(0f, root.getLayoutY(), 0.0f);
-    assertEquals(100f, root.getLayoutWidth(), 0.0f);
-    assertEquals(100f, root.getLayoutHeight(), 0.0f);
-
-    assertEquals(50f, root_child0.getLayoutX(), 0.0f);
-    assertEquals(0f, root_child0.getLayoutY(), 0.0f);
-    assertEquals(50f, root_child0.getLayoutWidth(), 0.0f);
-    assertEquals(50f, root_child0.getLayoutHeight(), 0.0f);
-
-    assertEquals(50f, root_child1.getLayoutX(), 0.0f);
-    assertEquals(50f, root_child1.getLayoutY(), 0.0f);
-    assertEquals(50f, root_child1.getLayoutWidth(), 0.0f);
-    assertEquals(20f, root_child1.getLayoutHeight(), 0.0f);
-  }
-
-  @Test
-  public void test_align_baseline_child_margin() {
-    final YogaNode root = new YogaNode();
-    root.setFlexDirection(YogaFlexDirection.ROW);
-    root.setAlignItems(YogaAlign.BASELINE);
-    root.setWidth(100f);
-    root.setHeight(100f);
-
-    final YogaNode root_child0 = new YogaNode();
-    root_child0.setMargin(YogaEdge.LEFT, 5f);
-    root_child0.setMargin(YogaEdge.TOP, 5f);
-    root_child0.setMargin(YogaEdge.RIGHT, 5f);
-    root_child0.setMargin(YogaEdge.BOTTOM, 5f);
-    root_child0.setWidth(50f);
-    root_child0.setHeight(50f);
-    root.addChildAt(root_child0, 0);
-
-    final YogaNode root_child1 = new YogaNode();
-    root_child1.setWidth(50f);
-    root_child1.setHeight(20f);
-    root.addChildAt(root_child1, 1);
-
-    final YogaNode root_child1_child0 = new YogaNode();
-    root_child1_child0.setMargin(YogaEdge.LEFT, 1f);
-    root_child1_child0.setMargin(YogaEdge.TOP, 1f);
-    root_child1_child0.setMargin(YogaEdge.RIGHT, 1f);
-    root_child1_child0.setMargin(YogaEdge.BOTTOM, 1f);
-    root_child1_child0.setWidth(50f);
-    root_child1_child0.setHeight(10f);
-    root_child1.addChildAt(root_child1_child0, 0);
-    root.setDirection(YogaDirection.LTR);
-    root.calculateLayout();
-
-    assertEquals(0f, root.getLayoutX(), 0.0f);
-    assertEquals(0f, root.getLayoutY(), 0.0f);
-    assertEquals(100f, root.getLayoutWidth(), 0.0f);
-    assertEquals(100f, root.getLayoutHeight(), 0.0f);
-
-    assertEquals(5f, root_child0.getLayoutX(), 0.0f);
-    assertEquals(5f, root_child0.getLayoutY(), 0.0f);
-    assertEquals(50f, root_child0.getLayoutWidth(), 0.0f);
-    assertEquals(50f, root_child0.getLayoutHeight(), 0.0f);
-
-    assertEquals(60f, root_child1.getLayoutX(), 0.0f);
-    assertEquals(44f, root_child1.getLayoutY(), 0.0f);
-    assertEquals(50f, root_child1.getLayoutWidth(), 0.0f);
-    assertEquals(20f, root_child1.getLayoutHeight(), 0.0f);
-
-    assertEquals(1f, root_child1_child0.getLayoutX(), 0.0f);
-    assertEquals(1f, root_child1_child0.getLayoutY(), 0.0f);
-    assertEquals(50f, root_child1_child0.getLayoutWidth(), 0.0f);
-    assertEquals(10f, root_child1_child0.getLayoutHeight(), 0.0f);
-
-    root.setDirection(YogaDirection.RTL);
-    root.calculateLayout();
-
-    assertEquals(0f, root.getLayoutX(), 0.0f);
-    assertEquals(0f, root.getLayoutY(), 0.0f);
-    assertEquals(100f, root.getLayoutWidth(), 0.0f);
-    assertEquals(100f, root.getLayoutHeight(), 0.0f);
-
-    assertEquals(45f, root_child0.getLayoutX(), 0.0f);
-    assertEquals(5f, root_child0.getLayoutY(), 0.0f);
-    assertEquals(50f, root_child0.getLayoutWidth(), 0.0f);
-    assertEquals(50f, root_child0.getLayoutHeight(), 0.0f);
-
-    assertEquals(-10f, root_child1.getLayoutX(), 0.0f);
-    assertEquals(44f, root_child1.getLayoutY(), 0.0f);
-    assertEquals(50f, root_child1.getLayoutWidth(), 0.0f);
-    assertEquals(20f, root_child1.getLayoutHeight(), 0.0f);
-
-    assertEquals(-1f, root_child1_child0.getLayoutX(), 0.0f);
-    assertEquals(1f, root_child1_child0.getLayoutY(), 0.0f);
-    assertEquals(50f, root_child1_child0.getLayoutWidth(), 0.0f);
-    assertEquals(10f, root_child1_child0.getLayoutHeight(), 0.0f);
-  }
-
-  @Test
-  public void test_align_baseline_multiline() {
-    final YogaNode root = new YogaNode();
-    root.setFlexDirection(YogaFlexDirection.ROW);
-    root.setAlignItems(YogaAlign.BASELINE);
-    root.setWrap(YogaWrap.WRAP);
-    root.setWidth(100f);
-    root.setHeight(100f);
-
-    final YogaNode root_child0 = new YogaNode();
-    root_child0.setWidth(50f);
-    root_child0.setHeight(50f);
-    root.addChildAt(root_child0, 0);
-
-    final YogaNode root_child1 = new YogaNode();
-    root_child1.setWidth(50f);
-    root_child1.setHeight(20f);
-    root.addChildAt(root_child1, 1);
-
-    final YogaNode root_child1_child0 = new YogaNode();
-    root_child1_child0.setWidth(50f);
-    root_child1_child0.setHeight(10f);
-    root_child1.addChildAt(root_child1_child0, 0);
-
-    final YogaNode root_child2 = new YogaNode();
-    root_child2.setWidth(50f);
-    root_child2.setHeight(20f);
-    root.addChildAt(root_child2, 2);
-
-    final YogaNode root_child2_child0 = new YogaNode();
-    root_child2_child0.setWidth(50f);
-    root_child2_child0.setHeight(10f);
-    root_child2.addChildAt(root_child2_child0, 0);
-
-    final YogaNode root_child3 = new YogaNode();
-    root_child3.setWidth(50f);
-    root_child3.setHeight(50f);
-    root.addChildAt(root_child3, 3);
-    root.setDirection(YogaDirection.LTR);
-    root.calculateLayout();
-
-    assertEquals(0f, root.getLayoutX(), 0.0f);
-    assertEquals(0f, root.getLayoutY(), 0.0f);
-    assertEquals(100f, root.getLayoutWidth(), 0.0f);
-    assertEquals(100f, root.getLayoutHeight(), 0.0f);
-
-    assertEquals(0f, root_child0.getLayoutX(), 0.0f);
-    assertEquals(0f, root_child0.getLayoutY(), 0.0f);
-    assertEquals(50f, root_child0.getLayoutWidth(), 0.0f);
-    assertEquals(50f, root_child0.getLayoutHeight(), 0.0f);
-
-    assertEquals(50f, root_child1.getLayoutX(), 0.0f);
-    assertEquals(40f, root_child1.getLayoutY(), 0.0f);
-    assertEquals(50f, root_child1.getLayoutWidth(), 0.0f);
-    assertEquals(20f, root_child1.getLayoutHeight(), 0.0f);
-
-    assertEquals(0f, root_child1_child0.getLayoutX(), 0.0f);
-    assertEquals(0f, root_child1_child0.getLayoutY(), 0.0f);
-    assertEquals(50f, root_child1_child0.getLayoutWidth(), 0.0f);
-    assertEquals(10f, root_child1_child0.getLayoutHeight(), 0.0f);
-
-    assertEquals(0f, root_child2.getLayoutX(), 0.0f);
-    assertEquals(100f, root_child2.getLayoutY(), 0.0f);
-    assertEquals(50f, root_child2.getLayoutWidth(), 0.0f);
-    assertEquals(20f, root_child2.getLayoutHeight(), 0.0f);
-
-    assertEquals(0f, root_child2_child0.getLayoutX(), 0.0f);
-    assertEquals(0f, root_child2_child0.getLayoutY(), 0.0f);
-    assertEquals(50f, root_child2_child0.getLayoutWidth(), 0.0f);
-    assertEquals(10f, root_child2_child0.getLayoutHeight(), 0.0f);
-
-    assertEquals(50f, root_child3.getLayoutX(), 0.0f);
-    assertEquals(60f, root_child3.getLayoutY(), 0.0f);
-    assertEquals(50f, root_child3.getLayoutWidth(), 0.0f);
-    assertEquals(50f, root_child3.getLayoutHeight(), 0.0f);
-
-    root.setDirection(YogaDirection.RTL);
-    root.calculateLayout();
-
-    assertEquals(0f, root.getLayoutX(), 0.0f);
-    assertEquals(0f, root.getLayoutY(), 0.0f);
-    assertEquals(100f, root.getLayoutWidth(), 0.0f);
-    assertEquals(100f, root.getLayoutHeight(), 0.0f);
-
-    assertEquals(50f, root_child0.getLayoutX(), 0.0f);
-    assertEquals(0f, root_child0.getLayoutY(), 0.0f);
-    assertEquals(50f, root_child0.getLayoutWidth(), 0.0f);
-    assertEquals(50f, root_child0.getLayoutHeight(), 0.0f);
-
-    assertEquals(0f, root_child1.getLayoutX(), 0.0f);
-    assertEquals(40f, root_child1.getLayoutY(), 0.0f);
-    assertEquals(50f, root_child1.getLayoutWidth(), 0.0f);
-    assertEquals(20f, root_child1.getLayoutHeight(), 0.0f);
-
-    assertEquals(0f, root_child1_child0.getLayoutX(), 0.0f);
-    assertEquals(0f, root_child1_child0.getLayoutY(), 0.0f);
-    assertEquals(50f, root_child1_child0.getLayoutWidth(), 0.0f);
-    assertEquals(10f, root_child1_child0.getLayoutHeight(), 0.0f);
-
-    assertEquals(50f, root_child2.getLayoutX(), 0.0f);
-    assertEquals(100f, root_child2.getLayoutY(), 0.0f);
-    assertEquals(50f, root_child2.getLayoutWidth(), 0.0f);
-    assertEquals(20f, root_child2.getLayoutHeight(), 0.0f);
-
-    assertEquals(0f, root_child2_child0.getLayoutX(), 0.0f);
-    assertEquals(0f, root_child2_child0.getLayoutY(), 0.0f);
-    assertEquals(50f, root_child2_child0.getLayoutWidth(), 0.0f);
-    assertEquals(10f, root_child2_child0.getLayoutHeight(), 0.0f);
-
-    assertEquals(0f, root_child3.getLayoutX(), 0.0f);
-    assertEquals(60f, root_child3.getLayoutY(), 0.0f);
-    assertEquals(50f, root_child3.getLayoutWidth(), 0.0f);
-    assertEquals(50f, root_child3.getLayoutHeight(), 0.0f);
-  }
-
-  @Test
-  public void test_align_baseline_self() {
-    final YogaNode root = new YogaNode();
-    root.setFlexDirection(YogaFlexDirection.ROW);
-    root.setWidth(100f);
-    root.setHeight(100f);
-
-    final YogaNode root_child0 = new YogaNode();
-    root_child0.setAlignSelf(YogaAlign.BASELINE);
-    root_child0.setWidth(50f);
-    root_child0.setHeight(50f);
-    root.addChildAt(root_child0, 0);
-
-    final YogaNode root_child1 = new YogaNode();
-    root_child1.setAlignSelf(YogaAlign.BASELINE);
-    root_child1.setWidth(50f);
-    root_child1.setHeight(20f);
-    root.addChildAt(root_child1, 1);
-
-    final YogaNode root_child1_child0 = new YogaNode();
-    root_child1_child0.setWidth(50f);
-    root_child1_child0.setHeight(10f);
-    root_child1.addChildAt(root_child1_child0, 0);
-    root.setDirection(YogaDirection.LTR);
-    root.calculateLayout();
-
-    assertEquals(0f, root.getLayoutX(), 0.0f);
-    assertEquals(0f, root.getLayoutY(), 0.0f);
-    assertEquals(100f, root.getLayoutWidth(), 0.0f);
-    assertEquals(100f, root.getLayoutHeight(), 0.0f);
-
-    assertEquals(0f, root_child0.getLayoutX(), 0.0f);
-    assertEquals(0f, root_child0.getLayoutY(), 0.0f);
-    assertEquals(50f, root_child0.getLayoutWidth(), 0.0f);
-    assertEquals(50f, root_child0.getLayoutHeight(), 0.0f);
-
-    assertEquals(50f, root_child1.getLayoutX(), 0.0f);
-    assertEquals(40f, root_child1.getLayoutY(), 0.0f);
-    assertEquals(50f, root_child1.getLayoutWidth(), 0.0f);
-    assertEquals(20f, root_child1.getLayoutHeight(), 0.0f);
-
-    assertEquals(0f, root_child1_child0.getLayoutX(), 0.0f);
-    assertEquals(0f, root_child1_child0.getLayoutY(), 0.0f);
-    assertEquals(50f, root_child1_child0.getLayoutWidth(), 0.0f);
-    assertEquals(10f, root_child1_child0.getLayoutHeight(), 0.0f);
-
-    root.setDirection(YogaDirection.RTL);
-    root.calculateLayout();
-
-    assertEquals(0f, root.getLayoutX(), 0.0f);
-    assertEquals(0f, root.getLayoutY(), 0.0f);
-    assertEquals(100f, root.getLayoutWidth(), 0.0f);
-    assertEquals(100f, root.getLayoutHeight(), 0.0f);
-
-    assertEquals(50f, root_child0.getLayoutX(), 0.0f);
-    assertEquals(0f, root_child0.getLayoutY(), 0.0f);
-    assertEquals(50f, root_child0.getLayoutWidth(), 0.0f);
-    assertEquals(50f, root_child0.getLayoutHeight(), 0.0f);
-
-    assertEquals(0f, root_child1.getLayoutX(), 0.0f);
-    assertEquals(40f, root_child1.getLayoutY(), 0.0f);
-    assertEquals(50f, root_child1.getLayoutWidth(), 0.0f);
-    assertEquals(20f, root_child1.getLayoutHeight(), 0.0f);
-
-    assertEquals(0f, root_child1_child0.getLayoutX(), 0.0f);
-    assertEquals(0f, root_child1_child0.getLayoutY(), 0.0f);
-    assertEquals(50f, root_child1_child0.getLayoutWidth(), 0.0f);
-    assertEquals(10f, root_child1_child0.getLayoutHeight(), 0.0f);
-  }
-
-  @Test
-  public void test_align_baseline_multiline_column() {
-    final YogaNode root = new YogaNode();
-    root.setAlignItems(YogaAlign.BASELINE);
-    root.setWrap(YogaWrap.WRAP);
-    root.setWidth(100f);
-    root.setHeight(100f);
-
-    final YogaNode root_child0 = new YogaNode();
-    root_child0.setWidth(50f);
-    root_child0.setHeight(50f);
-    root.addChildAt(root_child0, 0);
-
-    final YogaNode root_child1 = new YogaNode();
-    root_child1.setWidth(50f);
-    root_child1.setHeight(50f);
-    root.addChildAt(root_child1, 1);
-
-    final YogaNode root_child1_child0 = new YogaNode();
-    root_child1_child0.setWidth(50f);
-    root_child1_child0.setHeight(10f);
-    root_child1.addChildAt(root_child1_child0, 0);
-
-    final YogaNode root_child2 = new YogaNode();
-    root_child2.setWidth(50f);
-    root_child2.setHeight(20f);
-    root.addChildAt(root_child2, 2);
-
-    final YogaNode root_child2_child0 = new YogaNode();
-    root_child2_child0.setWidth(50f);
-    root_child2_child0.setHeight(10f);
-    root_child2.addChildAt(root_child2_child0, 0);
-
-    final YogaNode root_child3 = new YogaNode();
-    root_child3.setWidth(50f);
-    root_child3.setHeight(20f);
-    root.addChildAt(root_child3, 3);
-    root.setDirection(YogaDirection.LTR);
-    root.calculateLayout();
-
-    assertEquals(0f, root.getLayoutX(), 0.0f);
-    assertEquals(0f, root.getLayoutY(), 0.0f);
-    assertEquals(100f, root.getLayoutWidth(), 0.0f);
-    assertEquals(100f, root.getLayoutHeight(), 0.0f);
-
-    assertEquals(0f, root_child0.getLayoutX(), 0.0f);
-    assertEquals(0f, root_child0.getLayoutY(), 0.0f);
-    assertEquals(50f, root_child0.getLayoutWidth(), 0.0f);
-    assertEquals(50f, root_child0.getLayoutHeight(), 0.0f);
-
-    assertEquals(0f, root_child1.getLayoutX(), 0.0f);
-    assertEquals(50f, root_child1.getLayoutY(), 0.0f);
-    assertEquals(50f, root_child1.getLayoutWidth(), 0.0f);
-    assertEquals(50f, root_child1.getLayoutHeight(), 0.0f);
-
-    assertEquals(0f, root_child1_child0.getLayoutX(), 0.0f);
-    assertEquals(0f, root_child1_child0.getLayoutY(), 0.0f);
-    assertEquals(50f, root_child1_child0.getLayoutWidth(), 0.0f);
-    assertEquals(10f, root_child1_child0.getLayoutHeight(), 0.0f);
-
-    assertEquals(50f, root_child2.getLayoutX(), 0.0f);
-    assertEquals(0f, root_child2.getLayoutY(), 0.0f);
-    assertEquals(50f, root_child2.getLayoutWidth(), 0.0f);
-    assertEquals(20f, root_child2.getLayoutHeight(), 0.0f);
-
-    assertEquals(0f, root_child2_child0.getLayoutX(), 0.0f);
-    assertEquals(0f, root_child2_child0.getLayoutY(), 0.0f);
-    assertEquals(50f, root_child2_child0.getLayoutWidth(), 0.0f);
-    assertEquals(10f, root_child2_child0.getLayoutHeight(), 0.0f);
-
-    assertEquals(50f, root_child3.getLayoutX(), 0.0f);
-    assertEquals(20f, root_child3.getLayoutY(), 0.0f);
-    assertEquals(50f, root_child3.getLayoutWidth(), 0.0f);
-    assertEquals(20f, root_child3.getLayoutHeight(), 0.0f);
-
-    root.setDirection(YogaDirection.RTL);
-    root.calculateLayout();
-
-    assertEquals(0f, root.getLayoutX(), 0.0f);
-    assertEquals(0f, root.getLayoutY(), 0.0f);
-    assertEquals(100f, root.getLayoutWidth(), 0.0f);
-    assertEquals(100f, root.getLayoutHeight(), 0.0f);
-
-    assertEquals(50f, root_child0.getLayoutX(), 0.0f);
-    assertEquals(0f, root_child0.getLayoutY(), 0.0f);
-    assertEquals(50f, root_child0.getLayoutWidth(), 0.0f);
-    assertEquals(50f, root_child0.getLayoutHeight(), 0.0f);
-
-    assertEquals(50f, root_child1.getLayoutX(), 0.0f);
-    assertEquals(50f, root_child1.getLayoutY(), 0.0f);
-    assertEquals(50f, root_child1.getLayoutWidth(), 0.0f);
-    assertEquals(50f, root_child1.getLayoutHeight(), 0.0f);
-
-    assertEquals(0f, root_child1_child0.getLayoutX(), 0.0f);
-    assertEquals(0f, root_child1_child0.getLayoutY(), 0.0f);
-    assertEquals(50f, root_child1_child0.getLayoutWidth(), 0.0f);
-    assertEquals(10f, root_child1_child0.getLayoutHeight(), 0.0f);
-
-    assertEquals(0f, root_child2.getLayoutX(), 0.0f);
-    assertEquals(0f, root_child2.getLayoutY(), 0.0f);
-    assertEquals(50f, root_child2.getLayoutWidth(), 0.0f);
-    assertEquals(20f, root_child2.getLayoutHeight(), 0.0f);
-
-    assertEquals(0f, root_child2_child0.getLayoutX(), 0.0f);
-    assertEquals(0f, root_child2_child0.getLayoutY(), 0.0f);
-    assertEquals(50f, root_child2_child0.getLayoutWidth(), 0.0f);
-    assertEquals(10f, root_child2_child0.getLayoutHeight(), 0.0f);
-
-    assertEquals(0f, root_child3.getLayoutX(), 0.0f);
-    assertEquals(20f, root_child3.getLayoutY(), 0.0f);
-    assertEquals(50f, root_child3.getLayoutWidth(), 0.0f);
-    assertEquals(20f, root_child3.getLayoutHeight(), 0.0f);
-  }
-
-  @Test
-  public void test_align_baseline_multiline_row_and_column() {
-    final YogaNode root = new YogaNode();
-    root.setFlexDirection(YogaFlexDirection.ROW);
-    root.setAlignItems(YogaAlign.BASELINE);
-    root.setWrap(YogaWrap.WRAP);
-    root.setWidth(100f);
-    root.setHeight(100f);
-
-    final YogaNode root_child0 = new YogaNode();
-    root_child0.setWidth(50f);
-    root_child0.setHeight(50f);
-    root.addChildAt(root_child0, 0);
-
-    final YogaNode root_child1 = new YogaNode();
-    root_child1.setWidth(50f);
-    root_child1.setHeight(50f);
-    root.addChildAt(root_child1, 1);
-
-    final YogaNode root_child1_child0 = new YogaNode();
-    root_child1_child0.setWidth(50f);
-    root_child1_child0.setHeight(10f);
-    root_child1.addChildAt(root_child1_child0, 0);
-
-    final YogaNode root_child2 = new YogaNode();
-    root_child2.setWidth(50f);
-    root_child2.setHeight(20f);
-    root.addChildAt(root_child2, 2);
-
-    final YogaNode root_child2_child0 = new YogaNode();
-    root_child2_child0.setWidth(50f);
-    root_child2_child0.setHeight(10f);
-    root_child2.addChildAt(root_child2_child0, 0);
-
-    final YogaNode root_child3 = new YogaNode();
-    root_child3.setWidth(50f);
-    root_child3.setHeight(20f);
-    root.addChildAt(root_child3, 3);
-    root.setDirection(YogaDirection.LTR);
-    root.calculateLayout();
-
-    assertEquals(0f, root.getLayoutX(), 0.0f);
-    assertEquals(0f, root.getLayoutY(), 0.0f);
-    assertEquals(100f, root.getLayoutWidth(), 0.0f);
-    assertEquals(100f, root.getLayoutHeight(), 0.0f);
-
-    assertEquals(0f, root_child0.getLayoutX(), 0.0f);
-    assertEquals(0f, root_child0.getLayoutY(), 0.0f);
-    assertEquals(50f, root_child0.getLayoutWidth(), 0.0f);
-    assertEquals(50f, root_child0.getLayoutHeight(), 0.0f);
-
-    assertEquals(50f, root_child1.getLayoutX(), 0.0f);
-    assertEquals(40f, root_child1.getLayoutY(), 0.0f);
-    assertEquals(50f, root_child1.getLayoutWidth(), 0.0f);
-    assertEquals(50f, root_child1.getLayoutHeight(), 0.0f);
-
-    assertEquals(0f, root_child1_child0.getLayoutX(), 0.0f);
-    assertEquals(0f, root_child1_child0.getLayoutY(), 0.0f);
-    assertEquals(50f, root_child1_child0.getLayoutWidth(), 0.0f);
-    assertEquals(10f, root_child1_child0.getLayoutHeight(), 0.0f);
-
-    assertEquals(0f, root_child2.getLayoutX(), 0.0f);
-    assertEquals(100f, root_child2.getLayoutY(), 0.0f);
-    assertEquals(50f, root_child2.getLayoutWidth(), 0.0f);
-    assertEquals(20f, root_child2.getLayoutHeight(), 0.0f);
-
-    assertEquals(0f, root_child2_child0.getLayoutX(), 0.0f);
-    assertEquals(0f, root_child2_child0.getLayoutY(), 0.0f);
-    assertEquals(50f, root_child2_child0.getLayoutWidth(), 0.0f);
-    assertEquals(10f, root_child2_child0.getLayoutHeight(), 0.0f);
-
-    assertEquals(50f, root_child3.getLayoutX(), 0.0f);
-    assertEquals(90f, root_child3.getLayoutY(), 0.0f);
-    assertEquals(50f, root_child3.getLayoutWidth(), 0.0f);
-    assertEquals(20f, root_child3.getLayoutHeight(), 0.0f);
-
-    root.setDirection(YogaDirection.RTL);
-    root.calculateLayout();
-
-    assertEquals(0f, root.getLayoutX(), 0.0f);
-    assertEquals(0f, root.getLayoutY(), 0.0f);
-    assertEquals(100f, root.getLayoutWidth(), 0.0f);
-    assertEquals(100f, root.getLayoutHeight(), 0.0f);
-
-    assertEquals(50f, root_child0.getLayoutX(), 0.0f);
-    assertEquals(0f, root_child0.getLayoutY(), 0.0f);
-    assertEquals(50f, root_child0.getLayoutWidth(), 0.0f);
-    assertEquals(50f, root_child0.getLayoutHeight(), 0.0f);
-
-    assertEquals(0f, root_child1.getLayoutX(), 0.0f);
-    assertEquals(40f, root_child1.getLayoutY(), 0.0f);
-    assertEquals(50f, root_child1.getLayoutWidth(), 0.0f);
-    assertEquals(50f, root_child1.getLayoutHeight(), 0.0f);
-
-    assertEquals(0f, root_child1_child0.getLayoutX(), 0.0f);
-    assertEquals(0f, root_child1_child0.getLayoutY(), 0.0f);
-    assertEquals(50f, root_child1_child0.getLayoutWidth(), 0.0f);
-    assertEquals(10f, root_child1_child0.getLayoutHeight(), 0.0f);
-
-    assertEquals(50f, root_child2.getLayoutX(), 0.0f);
-    assertEquals(100f, root_child2.getLayoutY(), 0.0f);
-    assertEquals(50f, root_child2.getLayoutWidth(), 0.0f);
-    assertEquals(20f, root_child2.getLayoutHeight(), 0.0f);
-
-    assertEquals(0f, root_child2_child0.getLayoutX(), 0.0f);
-    assertEquals(0f, root_child2_child0.getLayoutY(), 0.0f);
-    assertEquals(50f, root_child2_child0.getLayoutWidth(), 0.0f);
-    assertEquals(10f, root_child2_child0.getLayoutHeight(), 0.0f);
-
-    assertEquals(0f, root_child3.getLayoutX(), 0.0f);
-    assertEquals(90f, root_child3.getLayoutY(), 0.0f);
-    assertEquals(50f, root_child3.getLayoutWidth(), 0.0f);
-    assertEquals(20f, root_child3.getLayoutHeight(), 0.0f);
-  }
-
-}
