--- tests/YGAlignBaseline.cpp
@@ -1,893 +1,0 @@
-/**
- * Copyright (c) 2014-present, Facebook, Inc.
- * All rights reserved.
- *
- * This source code is licensed under the BSD-style license found in the
- * LICENSE file in the root directory of this source tree. An additional grant
- * of patent rights can be found in the PATENTS file in the same directory.
- */
-
- // @Generated by gentest/gentest.rb from gentest/fixtures/YGAlignBaseline.html
-
-#include <yoga/Yoga.h>
-#include <gtest/gtest.h>
-
-TEST(YogaTest, align_baseline) {
-  const YGNodeRef root = YGNodeNew();
-  YGNodeStyleSetFlexDirection(root, YGFlexDirectionRow);
-  YGNodeStyleSetAlignItems(root, YGAlignBaseline);
-  YGNodeStyleSetWidth(root, 100);
-  YGNodeStyleSetHeight(root, 100);
-
-  const YGNodeRef root_child0 = YGNodeNew();
-  YGNodeStyleSetWidth(root_child0, 50);
-  YGNodeStyleSetHeight(root_child0, 50);
-  YGNodeInsertChild(root, root_child0, 0);
-
-  const YGNodeRef root_child1 = YGNodeNew();
-  YGNodeStyleSetWidth(root_child1, 50);
-  YGNodeStyleSetHeight(root_child1, 20);
-  YGNodeInsertChild(root, root_child1, 1);
-  YGNodeCalculateLayout(root, YGUndefined, YGUndefined, YGDirectionLTR);
-
-  ASSERT_FLOAT_EQ(0, YGNodeLayoutGetLeft(root));
-  ASSERT_FLOAT_EQ(0, YGNodeLayoutGetTop(root));
-  ASSERT_FLOAT_EQ(100, YGNodeLayoutGetWidth(root));
-  ASSERT_FLOAT_EQ(100, YGNodeLayoutGetHeight(root));
-
-  ASSERT_FLOAT_EQ(0, YGNodeLayoutGetLeft(root_child0));
-  ASSERT_FLOAT_EQ(0, YGNodeLayoutGetTop(root_child0));
-  ASSERT_FLOAT_EQ(50, YGNodeLayoutGetWidth(root_child0));
-  ASSERT_FLOAT_EQ(50, YGNodeLayoutGetHeight(root_child0));
-
-  ASSERT_FLOAT_EQ(50, YGNodeLayoutGetLeft(root_child1));
-  ASSERT_FLOAT_EQ(30, YGNodeLayoutGetTop(root_child1));
-  ASSERT_FLOAT_EQ(50, YGNodeLayoutGetWidth(root_child1));
-  ASSERT_FLOAT_EQ(20, YGNodeLayoutGetHeight(root_child1));
-
-  YGNodeCalculateLayout(root, YGUndefined, YGUndefined, YGDirectionRTL);
-
-  ASSERT_FLOAT_EQ(0, YGNodeLayoutGetLeft(root));
-  ASSERT_FLOAT_EQ(0, YGNodeLayoutGetTop(root));
-  ASSERT_FLOAT_EQ(100, YGNodeLayoutGetWidth(root));
-  ASSERT_FLOAT_EQ(100, YGNodeLayoutGetHeight(root));
-
-  ASSERT_FLOAT_EQ(50, YGNodeLayoutGetLeft(root_child0));
-  ASSERT_FLOAT_EQ(0, YGNodeLayoutGetTop(root_child0));
-  ASSERT_FLOAT_EQ(50, YGNodeLayoutGetWidth(root_child0));
-  ASSERT_FLOAT_EQ(50, YGNodeLayoutGetHeight(root_child0));
-
-  ASSERT_FLOAT_EQ(0, YGNodeLayoutGetLeft(root_child1));
-  ASSERT_FLOAT_EQ(30, YGNodeLayoutGetTop(root_child1));
-  ASSERT_FLOAT_EQ(50, YGNodeLayoutGetWidth(root_child1));
-  ASSERT_FLOAT_EQ(20, YGNodeLayoutGetHeight(root_child1));
-
-  YGNodeFreeRecursive(root);
-}
-
-TEST(YogaTest, align_baseline_child) {
-  const YGNodeRef root = YGNodeNew();
-  YGNodeStyleSetFlexDirection(root, YGFlexDirectionRow);
-  YGNodeStyleSetAlignItems(root, YGAlignBaseline);
-  YGNodeStyleSetWidth(root, 100);
-  YGNodeStyleSetHeight(root, 100);
-
-  const YGNodeRef root_child0 = YGNodeNew();
-  YGNodeStyleSetWidth(root_child0, 50);
-  YGNodeStyleSetHeight(root_child0, 50);
-  YGNodeInsertChild(root, root_child0, 0);
-
-  const YGNodeRef root_child1 = YGNodeNew();
-  YGNodeStyleSetWidth(root_child1, 50);
-  YGNodeStyleSetHeight(root_child1, 20);
-  YGNodeInsertChild(root, root_child1, 1);
-
-  const YGNodeRef root_child1_child0 = YGNodeNew();
-  YGNodeStyleSetWidth(root_child1_child0, 50);
-  YGNodeStyleSetHeight(root_child1_child0, 10);
-  YGNodeInsertChild(root_child1, root_child1_child0, 0);
-  YGNodeCalculateLayout(root, YGUndefined, YGUndefined, YGDirectionLTR);
-
-  ASSERT_FLOAT_EQ(0, YGNodeLayoutGetLeft(root));
-  ASSERT_FLOAT_EQ(0, YGNodeLayoutGetTop(root));
-  ASSERT_FLOAT_EQ(100, YGNodeLayoutGetWidth(root));
-  ASSERT_FLOAT_EQ(100, YGNodeLayoutGetHeight(root));
-
-  ASSERT_FLOAT_EQ(0, YGNodeLayoutGetLeft(root_child0));
-  ASSERT_FLOAT_EQ(0, YGNodeLayoutGetTop(root_child0));
-  ASSERT_FLOAT_EQ(50, YGNodeLayoutGetWidth(root_child0));
-  ASSERT_FLOAT_EQ(50, YGNodeLayoutGetHeight(root_child0));
-
-  ASSERT_FLOAT_EQ(50, YGNodeLayoutGetLeft(root_child1));
-  ASSERT_FLOAT_EQ(40, YGNodeLayoutGetTop(root_child1));
-  ASSERT_FLOAT_EQ(50, YGNodeLayoutGetWidth(root_child1));
-  ASSERT_FLOAT_EQ(20, YGNodeLayoutGetHeight(root_child1));
-
-  ASSERT_FLOAT_EQ(0, YGNodeLayoutGetLeft(root_child1_child0));
-  ASSERT_FLOAT_EQ(0, YGNodeLayoutGetTop(root_child1_child0));
-  ASSERT_FLOAT_EQ(50, YGNodeLayoutGetWidth(root_child1_child0));
-  ASSERT_FLOAT_EQ(10, YGNodeLayoutGetHeight(root_child1_child0));
-
-  YGNodeCalculateLayout(root, YGUndefined, YGUndefined, YGDirectionRTL);
-
-  ASSERT_FLOAT_EQ(0, YGNodeLayoutGetLeft(root));
-  ASSERT_FLOAT_EQ(0, YGNodeLayoutGetTop(root));
-  ASSERT_FLOAT_EQ(100, YGNodeLayoutGetWidth(root));
-  ASSERT_FLOAT_EQ(100, YGNodeLayoutGetHeight(root));
-
-  ASSERT_FLOAT_EQ(50, YGNodeLayoutGetLeft(root_child0));
-  ASSERT_FLOAT_EQ(0, YGNodeLayoutGetTop(root_child0));
-  ASSERT_FLOAT_EQ(50, YGNodeLayoutGetWidth(root_child0));
-  ASSERT_FLOAT_EQ(50, YGNodeLayoutGetHeight(root_child0));
-
-  ASSERT_FLOAT_EQ(0, YGNodeLayoutGetLeft(root_child1));
-  ASSERT_FLOAT_EQ(40, YGNodeLayoutGetTop(root_child1));
-  ASSERT_FLOAT_EQ(50, YGNodeLayoutGetWidth(root_child1));
-  ASSERT_FLOAT_EQ(20, YGNodeLayoutGetHeight(root_child1));
-
-  ASSERT_FLOAT_EQ(0, YGNodeLayoutGetLeft(root_child1_child0));
-  ASSERT_FLOAT_EQ(0, YGNodeLayoutGetTop(root_child1_child0));
-  ASSERT_FLOAT_EQ(50, YGNodeLayoutGetWidth(root_child1_child0));
-  ASSERT_FLOAT_EQ(10, YGNodeLayoutGetHeight(root_child1_child0));
-
-  YGNodeFreeRecursive(root);
-}
-
-TEST(YogaTest, align_baseline_child_top) {
-  const YGNodeRef root = YGNodeNew();
-  YGNodeStyleSetFlexDirection(root, YGFlexDirectionRow);
-  YGNodeStyleSetAlignItems(root, YGAlignBaseline);
-  YGNodeStyleSetWidth(root, 100);
-  YGNodeStyleSetHeight(root, 100);
-
-  const YGNodeRef root_child0 = YGNodeNew();
-  YGNodeStyleSetPosition(root_child0, YGEdgeTop, 10);
-  YGNodeStyleSetWidth(root_child0, 50);
-  YGNodeStyleSetHeight(root_child0, 50);
-  YGNodeInsertChild(root, root_child0, 0);
-
-  const YGNodeRef root_child1 = YGNodeNew();
-  YGNodeStyleSetWidth(root_child1, 50);
-  YGNodeStyleSetHeight(root_child1, 20);
-  YGNodeInsertChild(root, root_child1, 1);
-
-  const YGNodeRef root_child1_child0 = YGNodeNew();
-  YGNodeStyleSetWidth(root_child1_child0, 50);
-  YGNodeStyleSetHeight(root_child1_child0, 10);
-  YGNodeInsertChild(root_child1, root_child1_child0, 0);
-  YGNodeCalculateLayout(root, YGUndefined, YGUndefined, YGDirectionLTR);
-
-  ASSERT_FLOAT_EQ(0, YGNodeLayoutGetLeft(root));
-  ASSERT_FLOAT_EQ(0, YGNodeLayoutGetTop(root));
-  ASSERT_FLOAT_EQ(100, YGNodeLayoutGetWidth(root));
-  ASSERT_FLOAT_EQ(100, YGNodeLayoutGetHeight(root));
-
-  ASSERT_FLOAT_EQ(0, YGNodeLayoutGetLeft(root_child0));
-  ASSERT_FLOAT_EQ(10, YGNodeLayoutGetTop(root_child0));
-  ASSERT_FLOAT_EQ(50, YGNodeLayoutGetWidth(root_child0));
-  ASSERT_FLOAT_EQ(50, YGNodeLayoutGetHeight(root_child0));
-
-  ASSERT_FLOAT_EQ(50, YGNodeLayoutGetLeft(root_child1));
-  ASSERT_FLOAT_EQ(40, YGNodeLayoutGetTop(root_child1));
-  ASSERT_FLOAT_EQ(50, YGNodeLayoutGetWidth(root_child1));
-  ASSERT_FLOAT_EQ(20, YGNodeLayoutGetHeight(root_child1));
-
-  ASSERT_FLOAT_EQ(0, YGNodeLayoutGetLeft(root_child1_child0));
-  ASSERT_FLOAT_EQ(0, YGNodeLayoutGetTop(root_child1_child0));
-  ASSERT_FLOAT_EQ(50, YGNodeLayoutGetWidth(root_child1_child0));
-  ASSERT_FLOAT_EQ(10, YGNodeLayoutGetHeight(root_child1_child0));
-
-  YGNodeCalculateLayout(root, YGUndefined, YGUndefined, YGDirectionRTL);
-
-  ASSERT_FLOAT_EQ(0, YGNodeLayoutGetLeft(root));
-  ASSERT_FLOAT_EQ(0, YGNodeLayoutGetTop(root));
-  ASSERT_FLOAT_EQ(100, YGNodeLayoutGetWidth(root));
-  ASSERT_FLOAT_EQ(100, YGNodeLayoutGetHeight(root));
-
-  ASSERT_FLOAT_EQ(50, YGNodeLayoutGetLeft(root_child0));
-  ASSERT_FLOAT_EQ(10, YGNodeLayoutGetTop(root_child0));
-  ASSERT_FLOAT_EQ(50, YGNodeLayoutGetWidth(root_child0));
-  ASSERT_FLOAT_EQ(50, YGNodeLayoutGetHeight(root_child0));
-
-  ASSERT_FLOAT_EQ(0, YGNodeLayoutGetLeft(root_child1));
-  ASSERT_FLOAT_EQ(40, YGNodeLayoutGetTop(root_child1));
-  ASSERT_FLOAT_EQ(50, YGNodeLayoutGetWidth(root_child1));
-  ASSERT_FLOAT_EQ(20, YGNodeLayoutGetHeight(root_child1));
-
-  ASSERT_FLOAT_EQ(0, YGNodeLayoutGetLeft(root_child1_child0));
-  ASSERT_FLOAT_EQ(0, YGNodeLayoutGetTop(root_child1_child0));
-  ASSERT_FLOAT_EQ(50, YGNodeLayoutGetWidth(root_child1_child0));
-  ASSERT_FLOAT_EQ(10, YGNodeLayoutGetHeight(root_child1_child0));
-
-  YGNodeFreeRecursive(root);
-}
-
-TEST(YogaTest, align_baseline_child_top2) {
-  const YGNodeRef root = YGNodeNew();
-  YGNodeStyleSetFlexDirection(root, YGFlexDirectionRow);
-  YGNodeStyleSetAlignItems(root, YGAlignBaseline);
-  YGNodeStyleSetWidth(root, 100);
-  YGNodeStyleSetHeight(root, 100);
-
-  const YGNodeRef root_child0 = YGNodeNew();
-  YGNodeStyleSetWidth(root_child0, 50);
-  YGNodeStyleSetHeight(root_child0, 50);
-  YGNodeInsertChild(root, root_child0, 0);
-
-  const YGNodeRef root_child1 = YGNodeNew();
-  YGNodeStyleSetPosition(root_child1, YGEdgeTop, 5);
-  YGNodeStyleSetWidth(root_child1, 50);
-  YGNodeStyleSetHeight(root_child1, 20);
-  YGNodeInsertChild(root, root_child1, 1);
-
-  const YGNodeRef root_child1_child0 = YGNodeNew();
-  YGNodeStyleSetWidth(root_child1_child0, 50);
-  YGNodeStyleSetHeight(root_child1_child0, 10);
-  YGNodeInsertChild(root_child1, root_child1_child0, 0);
-  YGNodeCalculateLayout(root, YGUndefined, YGUndefined, YGDirectionLTR);
-
-  ASSERT_FLOAT_EQ(0, YGNodeLayoutGetLeft(root));
-  ASSERT_FLOAT_EQ(0, YGNodeLayoutGetTop(root));
-  ASSERT_FLOAT_EQ(100, YGNodeLayoutGetWidth(root));
-  ASSERT_FLOAT_EQ(100, YGNodeLayoutGetHeight(root));
-
-  ASSERT_FLOAT_EQ(0, YGNodeLayoutGetLeft(root_child0));
-  ASSERT_FLOAT_EQ(0, YGNodeLayoutGetTop(root_child0));
-  ASSERT_FLOAT_EQ(50, YGNodeLayoutGetWidth(root_child0));
-  ASSERT_FLOAT_EQ(50, YGNodeLayoutGetHeight(root_child0));
-
-  ASSERT_FLOAT_EQ(50, YGNodeLayoutGetLeft(root_child1));
-  ASSERT_FLOAT_EQ(45, YGNodeLayoutGetTop(root_child1));
-  ASSERT_FLOAT_EQ(50, YGNodeLayoutGetWidth(root_child1));
-  ASSERT_FLOAT_EQ(20, YGNodeLayoutGetHeight(root_child1));
-
-  ASSERT_FLOAT_EQ(0, YGNodeLayoutGetLeft(root_child1_child0));
-  ASSERT_FLOAT_EQ(0, YGNodeLayoutGetTop(root_child1_child0));
-  ASSERT_FLOAT_EQ(50, YGNodeLayoutGetWidth(root_child1_child0));
-  ASSERT_FLOAT_EQ(10, YGNodeLayoutGetHeight(root_child1_child0));
-
-  YGNodeCalculateLayout(root, YGUndefined, YGUndefined, YGDirectionRTL);
-
-  ASSERT_FLOAT_EQ(0, YGNodeLayoutGetLeft(root));
-  ASSERT_FLOAT_EQ(0, YGNodeLayoutGetTop(root));
-  ASSERT_FLOAT_EQ(100, YGNodeLayoutGetWidth(root));
-  ASSERT_FLOAT_EQ(100, YGNodeLayoutGetHeight(root));
-
-  ASSERT_FLOAT_EQ(50, YGNodeLayoutGetLeft(root_child0));
-  ASSERT_FLOAT_EQ(0, YGNodeLayoutGetTop(root_child0));
-  ASSERT_FLOAT_EQ(50, YGNodeLayoutGetWidth(root_child0));
-  ASSERT_FLOAT_EQ(50, YGNodeLayoutGetHeight(root_child0));
-
-  ASSERT_FLOAT_EQ(0, YGNodeLayoutGetLeft(root_child1));
-  ASSERT_FLOAT_EQ(45, YGNodeLayoutGetTop(root_child1));
-  ASSERT_FLOAT_EQ(50, YGNodeLayoutGetWidth(root_child1));
-  ASSERT_FLOAT_EQ(20, YGNodeLayoutGetHeight(root_child1));
-
-  ASSERT_FLOAT_EQ(0, YGNodeLayoutGetLeft(root_child1_child0));
-  ASSERT_FLOAT_EQ(0, YGNodeLayoutGetTop(root_child1_child0));
-  ASSERT_FLOAT_EQ(50, YGNodeLayoutGetWidth(root_child1_child0));
-  ASSERT_FLOAT_EQ(10, YGNodeLayoutGetHeight(root_child1_child0));
-
-  YGNodeFreeRecursive(root);
-}
-
-TEST(YogaTest, align_baseline_double_nested_child) {
-  const YGNodeRef root = YGNodeNew();
-  YGNodeStyleSetFlexDirection(root, YGFlexDirectionRow);
-  YGNodeStyleSetAlignItems(root, YGAlignBaseline);
-  YGNodeStyleSetWidth(root, 100);
-  YGNodeStyleSetHeight(root, 100);
-
-  const YGNodeRef root_child0 = YGNodeNew();
-  YGNodeStyleSetWidth(root_child0, 50);
-  YGNodeStyleSetHeight(root_child0, 50);
-  YGNodeInsertChild(root, root_child0, 0);
-
-  const YGNodeRef root_child0_child0 = YGNodeNew();
-  YGNodeStyleSetWidth(root_child0_child0, 50);
-  YGNodeStyleSetHeight(root_child0_child0, 20);
-  YGNodeInsertChild(root_child0, root_child0_child0, 0);
-
-  const YGNodeRef root_child1 = YGNodeNew();
-  YGNodeStyleSetWidth(root_child1, 50);
-  YGNodeStyleSetHeight(root_child1, 20);
-  YGNodeInsertChild(root, root_child1, 1);
-
-  const YGNodeRef root_child1_child0 = YGNodeNew();
-  YGNodeStyleSetWidth(root_child1_child0, 50);
-  YGNodeStyleSetHeight(root_child1_child0, 15);
-  YGNodeInsertChild(root_child1, root_child1_child0, 0);
-  YGNodeCalculateLayout(root, YGUndefined, YGUndefined, YGDirectionLTR);
-
-  ASSERT_FLOAT_EQ(0, YGNodeLayoutGetLeft(root));
-  ASSERT_FLOAT_EQ(0, YGNodeLayoutGetTop(root));
-  ASSERT_FLOAT_EQ(100, YGNodeLayoutGetWidth(root));
-  ASSERT_FLOAT_EQ(100, YGNodeLayoutGetHeight(root));
-
-  ASSERT_FLOAT_EQ(0, YGNodeLayoutGetLeft(root_child0));
-  ASSERT_FLOAT_EQ(0, YGNodeLayoutGetTop(root_child0));
-  ASSERT_FLOAT_EQ(50, YGNodeLayoutGetWidth(root_child0));
-  ASSERT_FLOAT_EQ(50, YGNodeLayoutGetHeight(root_child0));
-
-  ASSERT_FLOAT_EQ(0, YGNodeLayoutGetLeft(root_child0_child0));
-  ASSERT_FLOAT_EQ(0, YGNodeLayoutGetTop(root_child0_child0));
-  ASSERT_FLOAT_EQ(50, YGNodeLayoutGetWidth(root_child0_child0));
-  ASSERT_FLOAT_EQ(20, YGNodeLayoutGetHeight(root_child0_child0));
-
-  ASSERT_FLOAT_EQ(50, YGNodeLayoutGetLeft(root_child1));
-  ASSERT_FLOAT_EQ(5, YGNodeLayoutGetTop(root_child1));
-  ASSERT_FLOAT_EQ(50, YGNodeLayoutGetWidth(root_child1));
-  ASSERT_FLOAT_EQ(20, YGNodeLayoutGetHeight(root_child1));
-
-  ASSERT_FLOAT_EQ(0, YGNodeLayoutGetLeft(root_child1_child0));
-  ASSERT_FLOAT_EQ(0, YGNodeLayoutGetTop(root_child1_child0));
-  ASSERT_FLOAT_EQ(50, YGNodeLayoutGetWidth(root_child1_child0));
-  ASSERT_FLOAT_EQ(15, YGNodeLayoutGetHeight(root_child1_child0));
-
-  YGNodeCalculateLayout(root, YGUndefined, YGUndefined, YGDirectionRTL);
-
-  ASSERT_FLOAT_EQ(0, YGNodeLayoutGetLeft(root));
-  ASSERT_FLOAT_EQ(0, YGNodeLayoutGetTop(root));
-  ASSERT_FLOAT_EQ(100, YGNodeLayoutGetWidth(root));
-  ASSERT_FLOAT_EQ(100, YGNodeLayoutGetHeight(root));
-
-  ASSERT_FLOAT_EQ(50, YGNodeLayoutGetLeft(root_child0));
-  ASSERT_FLOAT_EQ(0, YGNodeLayoutGetTop(root_child0));
-  ASSERT_FLOAT_EQ(50, YGNodeLayoutGetWidth(root_child0));
-  ASSERT_FLOAT_EQ(50, YGNodeLayoutGetHeight(root_child0));
-
-  ASSERT_FLOAT_EQ(0, YGNodeLayoutGetLeft(root_child0_child0));
-  ASSERT_FLOAT_EQ(0, YGNodeLayoutGetTop(root_child0_child0));
-  ASSERT_FLOAT_EQ(50, YGNodeLayoutGetWidth(root_child0_child0));
-  ASSERT_FLOAT_EQ(20, YGNodeLayoutGetHeight(root_child0_child0));
-
-  ASSERT_FLOAT_EQ(0, YGNodeLayoutGetLeft(root_child1));
-  ASSERT_FLOAT_EQ(5, YGNodeLayoutGetTop(root_child1));
-  ASSERT_FLOAT_EQ(50, YGNodeLayoutGetWidth(root_child1));
-  ASSERT_FLOAT_EQ(20, YGNodeLayoutGetHeight(root_child1));
-
-  ASSERT_FLOAT_EQ(0, YGNodeLayoutGetLeft(root_child1_child0));
-  ASSERT_FLOAT_EQ(0, YGNodeLayoutGetTop(root_child1_child0));
-  ASSERT_FLOAT_EQ(50, YGNodeLayoutGetWidth(root_child1_child0));
-  ASSERT_FLOAT_EQ(15, YGNodeLayoutGetHeight(root_child1_child0));
-
-  YGNodeFreeRecursive(root);
-}
-
-TEST(YogaTest, align_baseline_column) {
-  const YGNodeRef root = YGNodeNew();
-  YGNodeStyleSetAlignItems(root, YGAlignBaseline);
-  YGNodeStyleSetWidth(root, 100);
-  YGNodeStyleSetHeight(root, 100);
-
-  const YGNodeRef root_child0 = YGNodeNew();
-  YGNodeStyleSetWidth(root_child0, 50);
-  YGNodeStyleSetHeight(root_child0, 50);
-  YGNodeInsertChild(root, root_child0, 0);
-
-  const YGNodeRef root_child1 = YGNodeNew();
-  YGNodeStyleSetWidth(root_child1, 50);
-  YGNodeStyleSetHeight(root_child1, 20);
-  YGNodeInsertChild(root, root_child1, 1);
-  YGNodeCalculateLayout(root, YGUndefined, YGUndefined, YGDirectionLTR);
-
-  ASSERT_FLOAT_EQ(0, YGNodeLayoutGetLeft(root));
-  ASSERT_FLOAT_EQ(0, YGNodeLayoutGetTop(root));
-  ASSERT_FLOAT_EQ(100, YGNodeLayoutGetWidth(root));
-  ASSERT_FLOAT_EQ(100, YGNodeLayoutGetHeight(root));
-
-  ASSERT_FLOAT_EQ(0, YGNodeLayoutGetLeft(root_child0));
-  ASSERT_FLOAT_EQ(0, YGNodeLayoutGetTop(root_child0));
-  ASSERT_FLOAT_EQ(50, YGNodeLayoutGetWidth(root_child0));
-  ASSERT_FLOAT_EQ(50, YGNodeLayoutGetHeight(root_child0));
-
-  ASSERT_FLOAT_EQ(0, YGNodeLayoutGetLeft(root_child1));
-  ASSERT_FLOAT_EQ(50, YGNodeLayoutGetTop(root_child1));
-  ASSERT_FLOAT_EQ(50, YGNodeLayoutGetWidth(root_child1));
-  ASSERT_FLOAT_EQ(20, YGNodeLayoutGetHeight(root_child1));
-
-  YGNodeCalculateLayout(root, YGUndefined, YGUndefined, YGDirectionRTL);
-
-  ASSERT_FLOAT_EQ(0, YGNodeLayoutGetLeft(root));
-  ASSERT_FLOAT_EQ(0, YGNodeLayoutGetTop(root));
-  ASSERT_FLOAT_EQ(100, YGNodeLayoutGetWidth(root));
-  ASSERT_FLOAT_EQ(100, YGNodeLayoutGetHeight(root));
-
-  ASSERT_FLOAT_EQ(50, YGNodeLayoutGetLeft(root_child0));
-  ASSERT_FLOAT_EQ(0, YGNodeLayoutGetTop(root_child0));
-  ASSERT_FLOAT_EQ(50, YGNodeLayoutGetWidth(root_child0));
-  ASSERT_FLOAT_EQ(50, YGNodeLayoutGetHeight(root_child0));
-
-  ASSERT_FLOAT_EQ(50, YGNodeLayoutGetLeft(root_child1));
-  ASSERT_FLOAT_EQ(50, YGNodeLayoutGetTop(root_child1));
-  ASSERT_FLOAT_EQ(50, YGNodeLayoutGetWidth(root_child1));
-  ASSERT_FLOAT_EQ(20, YGNodeLayoutGetHeight(root_child1));
-
-  YGNodeFreeRecursive(root);
-}
-
-TEST(YogaTest, align_baseline_child_margin) {
-  const YGNodeRef root = YGNodeNew();
-  YGNodeStyleSetFlexDirection(root, YGFlexDirectionRow);
-  YGNodeStyleSetAlignItems(root, YGAlignBaseline);
-  YGNodeStyleSetWidth(root, 100);
-  YGNodeStyleSetHeight(root, 100);
-
-  const YGNodeRef root_child0 = YGNodeNew();
-  YGNodeStyleSetMargin(root_child0, YGEdgeLeft, 5);
-  YGNodeStyleSetMargin(root_child0, YGEdgeTop, 5);
-  YGNodeStyleSetMargin(root_child0, YGEdgeRight, 5);
-  YGNodeStyleSetMargin(root_child0, YGEdgeBottom, 5);
-  YGNodeStyleSetWidth(root_child0, 50);
-  YGNodeStyleSetHeight(root_child0, 50);
-  YGNodeInsertChild(root, root_child0, 0);
-
-  const YGNodeRef root_child1 = YGNodeNew();
-  YGNodeStyleSetWidth(root_child1, 50);
-  YGNodeStyleSetHeight(root_child1, 20);
-  YGNodeInsertChild(root, root_child1, 1);
-
-  const YGNodeRef root_child1_child0 = YGNodeNew();
-  YGNodeStyleSetMargin(root_child1_child0, YGEdgeLeft, 1);
-  YGNodeStyleSetMargin(root_child1_child0, YGEdgeTop, 1);
-  YGNodeStyleSetMargin(root_child1_child0, YGEdgeRight, 1);
-  YGNodeStyleSetMargin(root_child1_child0, YGEdgeBottom, 1);
-  YGNodeStyleSetWidth(root_child1_child0, 50);
-  YGNodeStyleSetHeight(root_child1_child0, 10);
-  YGNodeInsertChild(root_child1, root_child1_child0, 0);
-  YGNodeCalculateLayout(root, YGUndefined, YGUndefined, YGDirectionLTR);
-
-  ASSERT_FLOAT_EQ(0, YGNodeLayoutGetLeft(root));
-  ASSERT_FLOAT_EQ(0, YGNodeLayoutGetTop(root));
-  ASSERT_FLOAT_EQ(100, YGNodeLayoutGetWidth(root));
-  ASSERT_FLOAT_EQ(100, YGNodeLayoutGetHeight(root));
-
-  ASSERT_FLOAT_EQ(5, YGNodeLayoutGetLeft(root_child0));
-  ASSERT_FLOAT_EQ(5, YGNodeLayoutGetTop(root_child0));
-  ASSERT_FLOAT_EQ(50, YGNodeLayoutGetWidth(root_child0));
-  ASSERT_FLOAT_EQ(50, YGNodeLayoutGetHeight(root_child0));
-
-  ASSERT_FLOAT_EQ(60, YGNodeLayoutGetLeft(root_child1));
-  ASSERT_FLOAT_EQ(44, YGNodeLayoutGetTop(root_child1));
-  ASSERT_FLOAT_EQ(50, YGNodeLayoutGetWidth(root_child1));
-  ASSERT_FLOAT_EQ(20, YGNodeLayoutGetHeight(root_child1));
-
-  ASSERT_FLOAT_EQ(1, YGNodeLayoutGetLeft(root_child1_child0));
-  ASSERT_FLOAT_EQ(1, YGNodeLayoutGetTop(root_child1_child0));
-  ASSERT_FLOAT_EQ(50, YGNodeLayoutGetWidth(root_child1_child0));
-  ASSERT_FLOAT_EQ(10, YGNodeLayoutGetHeight(root_child1_child0));
-
-  YGNodeCalculateLayout(root, YGUndefined, YGUndefined, YGDirectionRTL);
-
-  ASSERT_FLOAT_EQ(0, YGNodeLayoutGetLeft(root));
-  ASSERT_FLOAT_EQ(0, YGNodeLayoutGetTop(root));
-  ASSERT_FLOAT_EQ(100, YGNodeLayoutGetWidth(root));
-  ASSERT_FLOAT_EQ(100, YGNodeLayoutGetHeight(root));
-
-  ASSERT_FLOAT_EQ(45, YGNodeLayoutGetLeft(root_child0));
-  ASSERT_FLOAT_EQ(5, YGNodeLayoutGetTop(root_child0));
-  ASSERT_FLOAT_EQ(50, YGNodeLayoutGetWidth(root_child0));
-  ASSERT_FLOAT_EQ(50, YGNodeLayoutGetHeight(root_child0));
-
-  ASSERT_FLOAT_EQ(-10, YGNodeLayoutGetLeft(root_child1));
-  ASSERT_FLOAT_EQ(44, YGNodeLayoutGetTop(root_child1));
-  ASSERT_FLOAT_EQ(50, YGNodeLayoutGetWidth(root_child1));
-  ASSERT_FLOAT_EQ(20, YGNodeLayoutGetHeight(root_child1));
-
-  ASSERT_FLOAT_EQ(-1, YGNodeLayoutGetLeft(root_child1_child0));
-  ASSERT_FLOAT_EQ(1, YGNodeLayoutGetTop(root_child1_child0));
-  ASSERT_FLOAT_EQ(50, YGNodeLayoutGetWidth(root_child1_child0));
-  ASSERT_FLOAT_EQ(10, YGNodeLayoutGetHeight(root_child1_child0));
-
-  YGNodeFreeRecursive(root);
-}
-
-TEST(YogaTest, align_baseline_multiline) {
-  const YGNodeRef root = YGNodeNew();
-  YGNodeStyleSetFlexDirection(root, YGFlexDirectionRow);
-  YGNodeStyleSetAlignItems(root, YGAlignBaseline);
-  YGNodeStyleSetFlexWrap(root, YGWrapWrap);
-  YGNodeStyleSetWidth(root, 100);
-  YGNodeStyleSetHeight(root, 100);
-
-  const YGNodeRef root_child0 = YGNodeNew();
-  YGNodeStyleSetWidth(root_child0, 50);
-  YGNodeStyleSetHeight(root_child0, 50);
-  YGNodeInsertChild(root, root_child0, 0);
-
-  const YGNodeRef root_child1 = YGNodeNew();
-  YGNodeStyleSetWidth(root_child1, 50);
-  YGNodeStyleSetHeight(root_child1, 20);
-  YGNodeInsertChild(root, root_child1, 1);
-
-  const YGNodeRef root_child1_child0 = YGNodeNew();
-  YGNodeStyleSetWidth(root_child1_child0, 50);
-  YGNodeStyleSetHeight(root_child1_child0, 10);
-  YGNodeInsertChild(root_child1, root_child1_child0, 0);
-
-  const YGNodeRef root_child2 = YGNodeNew();
-  YGNodeStyleSetWidth(root_child2, 50);
-  YGNodeStyleSetHeight(root_child2, 20);
-  YGNodeInsertChild(root, root_child2, 2);
-
-  const YGNodeRef root_child2_child0 = YGNodeNew();
-  YGNodeStyleSetWidth(root_child2_child0, 50);
-  YGNodeStyleSetHeight(root_child2_child0, 10);
-  YGNodeInsertChild(root_child2, root_child2_child0, 0);
-
-  const YGNodeRef root_child3 = YGNodeNew();
-  YGNodeStyleSetWidth(root_child3, 50);
-  YGNodeStyleSetHeight(root_child3, 50);
-  YGNodeInsertChild(root, root_child3, 3);
-  YGNodeCalculateLayout(root, YGUndefined, YGUndefined, YGDirectionLTR);
-
-  ASSERT_FLOAT_EQ(0, YGNodeLayoutGetLeft(root));
-  ASSERT_FLOAT_EQ(0, YGNodeLayoutGetTop(root));
-  ASSERT_FLOAT_EQ(100, YGNodeLayoutGetWidth(root));
-  ASSERT_FLOAT_EQ(100, YGNodeLayoutGetHeight(root));
-
-  ASSERT_FLOAT_EQ(0, YGNodeLayoutGetLeft(root_child0));
-  ASSERT_FLOAT_EQ(0, YGNodeLayoutGetTop(root_child0));
-  ASSERT_FLOAT_EQ(50, YGNodeLayoutGetWidth(root_child0));
-  ASSERT_FLOAT_EQ(50, YGNodeLayoutGetHeight(root_child0));
-
-  ASSERT_FLOAT_EQ(50, YGNodeLayoutGetLeft(root_child1));
-  ASSERT_FLOAT_EQ(40, YGNodeLayoutGetTop(root_child1));
-  ASSERT_FLOAT_EQ(50, YGNodeLayoutGetWidth(root_child1));
-  ASSERT_FLOAT_EQ(20, YGNodeLayoutGetHeight(root_child1));
-
-  ASSERT_FLOAT_EQ(0, YGNodeLayoutGetLeft(root_child1_child0));
-  ASSERT_FLOAT_EQ(0, YGNodeLayoutGetTop(root_child1_child0));
-  ASSERT_FLOAT_EQ(50, YGNodeLayoutGetWidth(root_child1_child0));
-  ASSERT_FLOAT_EQ(10, YGNodeLayoutGetHeight(root_child1_child0));
-
-  ASSERT_FLOAT_EQ(0, YGNodeLayoutGetLeft(root_child2));
-  ASSERT_FLOAT_EQ(100, YGNodeLayoutGetTop(root_child2));
-  ASSERT_FLOAT_EQ(50, YGNodeLayoutGetWidth(root_child2));
-  ASSERT_FLOAT_EQ(20, YGNodeLayoutGetHeight(root_child2));
-
-  ASSERT_FLOAT_EQ(0, YGNodeLayoutGetLeft(root_child2_child0));
-  ASSERT_FLOAT_EQ(0, YGNodeLayoutGetTop(root_child2_child0));
-  ASSERT_FLOAT_EQ(50, YGNodeLayoutGetWidth(root_child2_child0));
-  ASSERT_FLOAT_EQ(10, YGNodeLayoutGetHeight(root_child2_child0));
-
-  ASSERT_FLOAT_EQ(50, YGNodeLayoutGetLeft(root_child3));
-  ASSERT_FLOAT_EQ(60, YGNodeLayoutGetTop(root_child3));
-  ASSERT_FLOAT_EQ(50, YGNodeLayoutGetWidth(root_child3));
-  ASSERT_FLOAT_EQ(50, YGNodeLayoutGetHeight(root_child3));
-
-  YGNodeCalculateLayout(root, YGUndefined, YGUndefined, YGDirectionRTL);
-
-  ASSERT_FLOAT_EQ(0, YGNodeLayoutGetLeft(root));
-  ASSERT_FLOAT_EQ(0, YGNodeLayoutGetTop(root));
-  ASSERT_FLOAT_EQ(100, YGNodeLayoutGetWidth(root));
-  ASSERT_FLOAT_EQ(100, YGNodeLayoutGetHeight(root));
-
-  ASSERT_FLOAT_EQ(50, YGNodeLayoutGetLeft(root_child0));
-  ASSERT_FLOAT_EQ(0, YGNodeLayoutGetTop(root_child0));
-  ASSERT_FLOAT_EQ(50, YGNodeLayoutGetWidth(root_child0));
-  ASSERT_FLOAT_EQ(50, YGNodeLayoutGetHeight(root_child0));
-
-  ASSERT_FLOAT_EQ(0, YGNodeLayoutGetLeft(root_child1));
-  ASSERT_FLOAT_EQ(40, YGNodeLayoutGetTop(root_child1));
-  ASSERT_FLOAT_EQ(50, YGNodeLayoutGetWidth(root_child1));
-  ASSERT_FLOAT_EQ(20, YGNodeLayoutGetHeight(root_child1));
-
-  ASSERT_FLOAT_EQ(0, YGNodeLayoutGetLeft(root_child1_child0));
-  ASSERT_FLOAT_EQ(0, YGNodeLayoutGetTop(root_child1_child0));
-  ASSERT_FLOAT_EQ(50, YGNodeLayoutGetWidth(root_child1_child0));
-  ASSERT_FLOAT_EQ(10, YGNodeLayoutGetHeight(root_child1_child0));
-
-  ASSERT_FLOAT_EQ(50, YGNodeLayoutGetLeft(root_child2));
-  ASSERT_FLOAT_EQ(100, YGNodeLayoutGetTop(root_child2));
-  ASSERT_FLOAT_EQ(50, YGNodeLayoutGetWidth(root_child2));
-  ASSERT_FLOAT_EQ(20, YGNodeLayoutGetHeight(root_child2));
-
-  ASSERT_FLOAT_EQ(0, YGNodeLayoutGetLeft(root_child2_child0));
-  ASSERT_FLOAT_EQ(0, YGNodeLayoutGetTop(root_child2_child0));
-  ASSERT_FLOAT_EQ(50, YGNodeLayoutGetWidth(root_child2_child0));
-  ASSERT_FLOAT_EQ(10, YGNodeLayoutGetHeight(root_child2_child0));
-
-  ASSERT_FLOAT_EQ(0, YGNodeLayoutGetLeft(root_child3));
-  ASSERT_FLOAT_EQ(60, YGNodeLayoutGetTop(root_child3));
-  ASSERT_FLOAT_EQ(50, YGNodeLayoutGetWidth(root_child3));
-  ASSERT_FLOAT_EQ(50, YGNodeLayoutGetHeight(root_child3));
-
-  YGNodeFreeRecursive(root);
-}
-
-TEST(YogaTest, align_baseline_self) {
-  const YGNodeRef root = YGNodeNew();
-  YGNodeStyleSetFlexDirection(root, YGFlexDirectionRow);
-  YGNodeStyleSetWidth(root, 100);
-  YGNodeStyleSetHeight(root, 100);
-
-  const YGNodeRef root_child0 = YGNodeNew();
-  YGNodeStyleSetAlignSelf(root_child0, YGAlignBaseline);
-  YGNodeStyleSetWidth(root_child0, 50);
-  YGNodeStyleSetHeight(root_child0, 50);
-  YGNodeInsertChild(root, root_child0, 0);
-
-  const YGNodeRef root_child1 = YGNodeNew();
-  YGNodeStyleSetAlignSelf(root_child1, YGAlignBaseline);
-  YGNodeStyleSetWidth(root_child1, 50);
-  YGNodeStyleSetHeight(root_child1, 20);
-  YGNodeInsertChild(root, root_child1, 1);
-
-  const YGNodeRef root_child1_child0 = YGNodeNew();
-  YGNodeStyleSetWidth(root_child1_child0, 50);
-  YGNodeStyleSetHeight(root_child1_child0, 10);
-  YGNodeInsertChild(root_child1, root_child1_child0, 0);
-  YGNodeCalculateLayout(root, YGUndefined, YGUndefined, YGDirectionLTR);
-
-  ASSERT_FLOAT_EQ(0, YGNodeLayoutGetLeft(root));
-  ASSERT_FLOAT_EQ(0, YGNodeLayoutGetTop(root));
-  ASSERT_FLOAT_EQ(100, YGNodeLayoutGetWidth(root));
-  ASSERT_FLOAT_EQ(100, YGNodeLayoutGetHeight(root));
-
-  ASSERT_FLOAT_EQ(0, YGNodeLayoutGetLeft(root_child0));
-  ASSERT_FLOAT_EQ(0, YGNodeLayoutGetTop(root_child0));
-  ASSERT_FLOAT_EQ(50, YGNodeLayoutGetWidth(root_child0));
-  ASSERT_FLOAT_EQ(50, YGNodeLayoutGetHeight(root_child0));
-
-  ASSERT_FLOAT_EQ(50, YGNodeLayoutGetLeft(root_child1));
-  ASSERT_FLOAT_EQ(40, YGNodeLayoutGetTop(root_child1));
-  ASSERT_FLOAT_EQ(50, YGNodeLayoutGetWidth(root_child1));
-  ASSERT_FLOAT_EQ(20, YGNodeLayoutGetHeight(root_child1));
-
-  ASSERT_FLOAT_EQ(0, YGNodeLayoutGetLeft(root_child1_child0));
-  ASSERT_FLOAT_EQ(0, YGNodeLayoutGetTop(root_child1_child0));
-  ASSERT_FLOAT_EQ(50, YGNodeLayoutGetWidth(root_child1_child0));
-  ASSERT_FLOAT_EQ(10, YGNodeLayoutGetHeight(root_child1_child0));
-
-  YGNodeCalculateLayout(root, YGUndefined, YGUndefined, YGDirectionRTL);
-
-  ASSERT_FLOAT_EQ(0, YGNodeLayoutGetLeft(root));
-  ASSERT_FLOAT_EQ(0, YGNodeLayoutGetTop(root));
-  ASSERT_FLOAT_EQ(100, YGNodeLayoutGetWidth(root));
-  ASSERT_FLOAT_EQ(100, YGNodeLayoutGetHeight(root));
-
-  ASSERT_FLOAT_EQ(50, YGNodeLayoutGetLeft(root_child0));
-  ASSERT_FLOAT_EQ(0, YGNodeLayoutGetTop(root_child0));
-  ASSERT_FLOAT_EQ(50, YGNodeLayoutGetWidth(root_child0));
-  ASSERT_FLOAT_EQ(50, YGNodeLayoutGetHeight(root_child0));
-
-  ASSERT_FLOAT_EQ(0, YGNodeLayoutGetLeft(root_child1));
-  ASSERT_FLOAT_EQ(40, YGNodeLayoutGetTop(root_child1));
-  ASSERT_FLOAT_EQ(50, YGNodeLayoutGetWidth(root_child1));
-  ASSERT_FLOAT_EQ(20, YGNodeLayoutGetHeight(root_child1));
-
-  ASSERT_FLOAT_EQ(0, YGNodeLayoutGetLeft(root_child1_child0));
-  ASSERT_FLOAT_EQ(0, YGNodeLayoutGetTop(root_child1_child0));
-  ASSERT_FLOAT_EQ(50, YGNodeLayoutGetWidth(root_child1_child0));
-  ASSERT_FLOAT_EQ(10, YGNodeLayoutGetHeight(root_child1_child0));
-
-  YGNodeFreeRecursive(root);
-}
-
-TEST(YogaTest, align_baseline_multiline_column) {
-  const YGNodeRef root = YGNodeNew();
-  YGNodeStyleSetAlignItems(root, YGAlignBaseline);
-  YGNodeStyleSetFlexWrap(root, YGWrapWrap);
-  YGNodeStyleSetWidth(root, 100);
-  YGNodeStyleSetHeight(root, 100);
-
-  const YGNodeRef root_child0 = YGNodeNew();
-  YGNodeStyleSetWidth(root_child0, 50);
-  YGNodeStyleSetHeight(root_child0, 50);
-  YGNodeInsertChild(root, root_child0, 0);
-
-  const YGNodeRef root_child1 = YGNodeNew();
-  YGNodeStyleSetWidth(root_child1, 50);
-  YGNodeStyleSetHeight(root_child1, 50);
-  YGNodeInsertChild(root, root_child1, 1);
-
-  const YGNodeRef root_child1_child0 = YGNodeNew();
-  YGNodeStyleSetWidth(root_child1_child0, 50);
-  YGNodeStyleSetHeight(root_child1_child0, 10);
-  YGNodeInsertChild(root_child1, root_child1_child0, 0);
-
-  const YGNodeRef root_child2 = YGNodeNew();
-  YGNodeStyleSetWidth(root_child2, 50);
-  YGNodeStyleSetHeight(root_child2, 20);
-  YGNodeInsertChild(root, root_child2, 2);
-
-  const YGNodeRef root_child2_child0 = YGNodeNew();
-  YGNodeStyleSetWidth(root_child2_child0, 50);
-  YGNodeStyleSetHeight(root_child2_child0, 10);
-  YGNodeInsertChild(root_child2, root_child2_child0, 0);
-
-  const YGNodeRef root_child3 = YGNodeNew();
-  YGNodeStyleSetWidth(root_child3, 50);
-  YGNodeStyleSetHeight(root_child3, 20);
-  YGNodeInsertChild(root, root_child3, 3);
-  YGNodeCalculateLayout(root, YGUndefined, YGUndefined, YGDirectionLTR);
-
-  ASSERT_FLOAT_EQ(0, YGNodeLayoutGetLeft(root));
-  ASSERT_FLOAT_EQ(0, YGNodeLayoutGetTop(root));
-  ASSERT_FLOAT_EQ(100, YGNodeLayoutGetWidth(root));
-  ASSERT_FLOAT_EQ(100, YGNodeLayoutGetHeight(root));
-
-  ASSERT_FLOAT_EQ(0, YGNodeLayoutGetLeft(root_child0));
-  ASSERT_FLOAT_EQ(0, YGNodeLayoutGetTop(root_child0));
-  ASSERT_FLOAT_EQ(50, YGNodeLayoutGetWidth(root_child0));
-  ASSERT_FLOAT_EQ(50, YGNodeLayoutGetHeight(root_child0));
-
-  ASSERT_FLOAT_EQ(0, YGNodeLayoutGetLeft(root_child1));
-  ASSERT_FLOAT_EQ(50, YGNodeLayoutGetTop(root_child1));
-  ASSERT_FLOAT_EQ(50, YGNodeLayoutGetWidth(root_child1));
-  ASSERT_FLOAT_EQ(50, YGNodeLayoutGetHeight(root_child1));
-
-  ASSERT_FLOAT_EQ(0, YGNodeLayoutGetLeft(root_child1_child0));
-  ASSERT_FLOAT_EQ(0, YGNodeLayoutGetTop(root_child1_child0));
-  ASSERT_FLOAT_EQ(50, YGNodeLayoutGetWidth(root_child1_child0));
-  ASSERT_FLOAT_EQ(10, YGNodeLayoutGetHeight(root_child1_child0));
-
-  ASSERT_FLOAT_EQ(50, YGNodeLayoutGetLeft(root_child2));
-  ASSERT_FLOAT_EQ(0, YGNodeLayoutGetTop(root_child2));
-  ASSERT_FLOAT_EQ(50, YGNodeLayoutGetWidth(root_child2));
-  ASSERT_FLOAT_EQ(20, YGNodeLayoutGetHeight(root_child2));
-
-  ASSERT_FLOAT_EQ(0, YGNodeLayoutGetLeft(root_child2_child0));
-  ASSERT_FLOAT_EQ(0, YGNodeLayoutGetTop(root_child2_child0));
-  ASSERT_FLOAT_EQ(50, YGNodeLayoutGetWidth(root_child2_child0));
-  ASSERT_FLOAT_EQ(10, YGNodeLayoutGetHeight(root_child2_child0));
-
-  ASSERT_FLOAT_EQ(50, YGNodeLayoutGetLeft(root_child3));
-  ASSERT_FLOAT_EQ(20, YGNodeLayoutGetTop(root_child3));
-  ASSERT_FLOAT_EQ(50, YGNodeLayoutGetWidth(root_child3));
-  ASSERT_FLOAT_EQ(20, YGNodeLayoutGetHeight(root_child3));
-
-  YGNodeCalculateLayout(root, YGUndefined, YGUndefined, YGDirectionRTL);
-
-  ASSERT_FLOAT_EQ(0, YGNodeLayoutGetLeft(root));
-  ASSERT_FLOAT_EQ(0, YGNodeLayoutGetTop(root));
-  ASSERT_FLOAT_EQ(100, YGNodeLayoutGetWidth(root));
-  ASSERT_FLOAT_EQ(100, YGNodeLayoutGetHeight(root));
-
-  ASSERT_FLOAT_EQ(50, YGNodeLayoutGetLeft(root_child0));
-  ASSERT_FLOAT_EQ(0, YGNodeLayoutGetTop(root_child0));
-  ASSERT_FLOAT_EQ(50, YGNodeLayoutGetWidth(root_child0));
-  ASSERT_FLOAT_EQ(50, YGNodeLayoutGetHeight(root_child0));
-
-  ASSERT_FLOAT_EQ(50, YGNodeLayoutGetLeft(root_child1));
-  ASSERT_FLOAT_EQ(50, YGNodeLayoutGetTop(root_child1));
-  ASSERT_FLOAT_EQ(50, YGNodeLayoutGetWidth(root_child1));
-  ASSERT_FLOAT_EQ(50, YGNodeLayoutGetHeight(root_child1));
-
-  ASSERT_FLOAT_EQ(0, YGNodeLayoutGetLeft(root_child1_child0));
-  ASSERT_FLOAT_EQ(0, YGNodeLayoutGetTop(root_child1_child0));
-  ASSERT_FLOAT_EQ(50, YGNodeLayoutGetWidth(root_child1_child0));
-  ASSERT_FLOAT_EQ(10, YGNodeLayoutGetHeight(root_child1_child0));
-
-  ASSERT_FLOAT_EQ(0, YGNodeLayoutGetLeft(root_child2));
-  ASSERT_FLOAT_EQ(0, YGNodeLayoutGetTop(root_child2));
-  ASSERT_FLOAT_EQ(50, YGNodeLayoutGetWidth(root_child2));
-  ASSERT_FLOAT_EQ(20, YGNodeLayoutGetHeight(root_child2));
-
-  ASSERT_FLOAT_EQ(0, YGNodeLayoutGetLeft(root_child2_child0));
-  ASSERT_FLOAT_EQ(0, YGNodeLayoutGetTop(root_child2_child0));
-  ASSERT_FLOAT_EQ(50, YGNodeLayoutGetWidth(root_child2_child0));
-  ASSERT_FLOAT_EQ(10, YGNodeLayoutGetHeight(root_child2_child0));
-
-  ASSERT_FLOAT_EQ(0, YGNodeLayoutGetLeft(root_child3));
-  ASSERT_FLOAT_EQ(20, YGNodeLayoutGetTop(root_child3));
-  ASSERT_FLOAT_EQ(50, YGNodeLayoutGetWidth(root_child3));
-  ASSERT_FLOAT_EQ(20, YGNodeLayoutGetHeight(root_child3));
-
-  YGNodeFreeRecursive(root);
-}
-
-TEST(YogaTest, align_baseline_multiline_row_and_column) {
-  const YGNodeRef root = YGNodeNew();
-  YGNodeStyleSetFlexDirection(root, YGFlexDirectionRow);
-  YGNodeStyleSetAlignItems(root, YGAlignBaseline);
-  YGNodeStyleSetFlexWrap(root, YGWrapWrap);
-  YGNodeStyleSetWidth(root, 100);
-  YGNodeStyleSetHeight(root, 100);
-
-  const YGNodeRef root_child0 = YGNodeNew();
-  YGNodeStyleSetWidth(root_child0, 50);
-  YGNodeStyleSetHeight(root_child0, 50);
-  YGNodeInsertChild(root, root_child0, 0);
-
-  const YGNodeRef root_child1 = YGNodeNew();
-  YGNodeStyleSetWidth(root_child1, 50);
-  YGNodeStyleSetHeight(root_child1, 50);
-  YGNodeInsertChild(root, root_child1, 1);
-
-  const YGNodeRef root_child1_child0 = YGNodeNew();
-  YGNodeStyleSetWidth(root_child1_child0, 50);
-  YGNodeStyleSetHeight(root_child1_child0, 10);
-  YGNodeInsertChild(root_child1, root_child1_child0, 0);
-
-  const YGNodeRef root_child2 = YGNodeNew();
-  YGNodeStyleSetWidth(root_child2, 50);
-  YGNodeStyleSetHeight(root_child2, 20);
-  YGNodeInsertChild(root, root_child2, 2);
-
-  const YGNodeRef root_child2_child0 = YGNodeNew();
-  YGNodeStyleSetWidth(root_child2_child0, 50);
-  YGNodeStyleSetHeight(root_child2_child0, 10);
-  YGNodeInsertChild(root_child2, root_child2_child0, 0);
-
-  const YGNodeRef root_child3 = YGNodeNew();
-  YGNodeStyleSetWidth(root_child3, 50);
-  YGNodeStyleSetHeight(root_child3, 20);
-  YGNodeInsertChild(root, root_child3, 3);
-  YGNodeCalculateLayout(root, YGUndefined, YGUndefined, YGDirectionLTR);
-
-  ASSERT_FLOAT_EQ(0, YGNodeLayoutGetLeft(root));
-  ASSERT_FLOAT_EQ(0, YGNodeLayoutGetTop(root));
-  ASSERT_FLOAT_EQ(100, YGNodeLayoutGetWidth(root));
-  ASSERT_FLOAT_EQ(100, YGNodeLayoutGetHeight(root));
-
-  ASSERT_FLOAT_EQ(0, YGNodeLayoutGetLeft(root_child0));
-  ASSERT_FLOAT_EQ(0, YGNodeLayoutGetTop(root_child0));
-  ASSERT_FLOAT_EQ(50, YGNodeLayoutGetWidth(root_child0));
-  ASSERT_FLOAT_EQ(50, YGNodeLayoutGetHeight(root_child0));
-
-  ASSERT_FLOAT_EQ(50, YGNodeLayoutGetLeft(root_child1));
-  ASSERT_FLOAT_EQ(40, YGNodeLayoutGetTop(root_child1));
-  ASSERT_FLOAT_EQ(50, YGNodeLayoutGetWidth(root_child1));
-  ASSERT_FLOAT_EQ(50, YGNodeLayoutGetHeight(root_child1));
-
-  ASSERT_FLOAT_EQ(0, YGNodeLayoutGetLeft(root_child1_child0));
-  ASSERT_FLOAT_EQ(0, YGNodeLayoutGetTop(root_child1_child0));
-  ASSERT_FLOAT_EQ(50, YGNodeLayoutGetWidth(root_child1_child0));
-  ASSERT_FLOAT_EQ(10, YGNodeLayoutGetHeight(root_child1_child0));
-
-  ASSERT_FLOAT_EQ(0, YGNodeLayoutGetLeft(root_child2));
-  ASSERT_FLOAT_EQ(100, YGNodeLayoutGetTop(root_child2));
-  ASSERT_FLOAT_EQ(50, YGNodeLayoutGetWidth(root_child2));
-  ASSERT_FLOAT_EQ(20, YGNodeLayoutGetHeight(root_child2));
-
-  ASSERT_FLOAT_EQ(0, YGNodeLayoutGetLeft(root_child2_child0));
-  ASSERT_FLOAT_EQ(0, YGNodeLayoutGetTop(root_child2_child0));
-  ASSERT_FLOAT_EQ(50, YGNodeLayoutGetWidth(root_child2_child0));
-  ASSERT_FLOAT_EQ(10, YGNodeLayoutGetHeight(root_child2_child0));
-
-  ASSERT_FLOAT_EQ(50, YGNodeLayoutGetLeft(root_child3));
-  ASSERT_FLOAT_EQ(90, YGNodeLayoutGetTop(root_child3));
-  ASSERT_FLOAT_EQ(50, YGNodeLayoutGetWidth(root_child3));
-  ASSERT_FLOAT_EQ(20, YGNodeLayoutGetHeight(root_child3));
-
-  YGNodeCalculateLayout(root, YGUndefined, YGUndefined, YGDirectionRTL);
-
-  ASSERT_FLOAT_EQ(0, YGNodeLayoutGetLeft(root));
-  ASSERT_FLOAT_EQ(0, YGNodeLayoutGetTop(root));
-  ASSERT_FLOAT_EQ(100, YGNodeLayoutGetWidth(root));
-  ASSERT_FLOAT_EQ(100, YGNodeLayoutGetHeight(root));
-
-  ASSERT_FLOAT_EQ(50, YGNodeLayoutGetLeft(root_child0));
-  ASSERT_FLOAT_EQ(0, YGNodeLayoutGetTop(root_child0));
-  ASSERT_FLOAT_EQ(50, YGNodeLayoutGetWidth(root_child0));
-  ASSERT_FLOAT_EQ(50, YGNodeLayoutGetHeight(root_child0));
-
-  ASSERT_FLOAT_EQ(0, YGNodeLayoutGetLeft(root_child1));
-  ASSERT_FLOAT_EQ(40, YGNodeLayoutGetTop(root_child1));
-  ASSERT_FLOAT_EQ(50, YGNodeLayoutGetWidth(root_child1));
-  ASSERT_FLOAT_EQ(50, YGNodeLayoutGetHeight(root_child1));
-
-  ASSERT_FLOAT_EQ(0, YGNodeLayoutGetLeft(root_child1_child0));
-  ASSERT_FLOAT_EQ(0, YGNodeLayoutGetTop(root_child1_child0));
-  ASSERT_FLOAT_EQ(50, YGNodeLayoutGetWidth(root_child1_child0));
-  ASSERT_FLOAT_EQ(10, YGNodeLayoutGetHeight(root_child1_child0));
-
-  ASSERT_FLOAT_EQ(50, YGNodeLayoutGetLeft(root_child2));
-  ASSERT_FLOAT_EQ(100, YGNodeLayoutGetTop(root_child2));
-  ASSERT_FLOAT_EQ(50, YGNodeLayoutGetWidth(root_child2));
-  ASSERT_FLOAT_EQ(20, YGNodeLayoutGetHeight(root_child2));
-
-  ASSERT_FLOAT_EQ(0, YGNodeLayoutGetLeft(root_child2_child0));
-  ASSERT_FLOAT_EQ(0, YGNodeLayoutGetTop(root_child2_child0));
-  ASSERT_FLOAT_EQ(50, YGNodeLayoutGetWidth(root_child2_child0));
-  ASSERT_FLOAT_EQ(10, YGNodeLayoutGetHeight(root_child2_child0));
-
-  ASSERT_FLOAT_EQ(0, YGNodeLayoutGetLeft(root_child3));
-  ASSERT_FLOAT_EQ(90, YGNodeLayoutGetTop(root_child3));
-  ASSERT_FLOAT_EQ(50, YGNodeLayoutGetWidth(root_child3));
-  ASSERT_FLOAT_EQ(20, YGNodeLayoutGetHeight(root_child3));
-
-  YGNodeFreeRecursive(root);
-}
